### 2019-08-21 #1668 @whedon
[PRE REVIEW]: The Walrus: the fastest calculation of hafnians, Hermite polynomials and Gaussian boson sampling
**Submitting author:** @nquesada (<a href="http://orcid.org/0000-0002-0175-1688">Nicolás Quesada</a>)
**Repository:** <a href="https://github.com/xanaduAI/thewalrus" target ="_blank">https://github.com/xanaduAI/thewalrus</a>
**Version:** 0.7.0
