### 2019-08-21 #1668 @nquesada
Hey @kyleniemeyer thanks for helping out in finding an editor for our submission!
I assume I can post here the people I can think could be helpful in reviewing, correct?
