### 2019-08-21 #1668 @nquesada
Great! From the list of potential reviewers JOSS has, I think either @goerz or @PhilipVinc would make good reviewers given their background in quantum optics.
