### 2019-08-22 05a320b
Adds tests for total_photon_num_dist_pure_state
