### 2019-08-22 2dd3701
Removes unnecesary else and unused argument
