### 2019-08-23 2141341
Update paper.bib
Updates bibliography in paper.bib
