+++ paper.bib
@@ -160,7 +160,7 @@
   author={Quesada, Nicolas and Arrazola, Juan Miguel},
   title="The Classical Complexity of Gaussian Boson Sampling",
   year={2019},
-  journal={in preparation}
+  journal={arXiv preprint 1908.08068}
 }
 
 
