### 2019-08-23 3e3a639
Apply suggestions from code review
Co-Authored-By: Josh Izaac <josh146@gmail.com>
