### 2019-08-23 e2a362b
Renames nmax to cutoff and parametrizes the tests
