+++ thewalrus/tests/test_quantum.py
@@ -677,15 +677,15 @@ def test_is_classical_cov_thermal(nbar):
     cov = (2 * nbar + 1) * np.identity(2)
     assert is_classical_cov(cov)
 
-
-def test_total_photon_num_dist_pure_state():
+@pytest.mark.parametrize("cutoff", [50, 51, 52, 53])
+def test_total_photon_num_dist_pure_state(cutoff):
     """ Test the correct photon number distribution is obtained for n modes
     with nmean number of photons up to Fock cutoff nmax"""
     n = 3
     nmean = 1.0
-    nmax = 50
+    cutoff = 50
     rs = np.arcsinh(np.sqrt(nmean)) * np.ones([n])
     cov = np.diag(np.concatenate([np.exp(2 * rs), np.exp(-2 * rs)]))
-    p1 = total_photon_num_dist_pure_state(cov, nmax=nmax)
-    p2 = gen_single_mode_dist(np.arcsinh(np.sqrt(nmean)), N=n, nmax=nmax)
+    p1 = total_photon_num_dist_pure_state(cov, cutoff=cutoff)
+    p2 = gen_single_mode_dist(np.arcsinh(np.sqrt(nmean)), N=n, cutoff=cutoff)
     assert np.allclose(p1, p2)
