### 2019-08-26 1786743
First tests. Still there is a problem in finding the right indices when calling density_matrix with no postselection
