### 2019-08-26 e42e9d8
Update paper.bib
