+++ paper.bib
@@ -1,6 +1,6 @@
 
 @article{caianiello1953,
-  title={On quantum field theory—I: explicit solution of Dyson’s equation in electrodynamics without use of feynman graphs},
+  title={On quantum field theory—I: explicit solution of {D}yson’s equation in electrodynamics without use of {F}eynman graphs},
   author={Caianiello, Eduardo R},
   journal={Il Nuovo Cimento (1943-1954)},
   volume={10},
@@ -44,9 +44,9 @@
 
 
 @inproceedings{bjorklund2012,
-  title={Counting perfect matchings as fast as Ryser},
+  title={Counting perfect matchings as fast as {R}yser},
   author={Bj{\"o}rklund, Andreas},
-  booktitle={Proceedings of the twenty-third annual ACM-SIAM symposium on Discrete Algorithms},
+  booktitle={Proceedings of the twenty-third annual {ACM-SIAM} symposium on Discrete Algorithms},
   pages={914--921},
   year={2012}
 }
@@ -55,7 +55,7 @@
 
 
 @article{kruse2018,
-  title={A detailed study of Gaussian Boson Sampling},
+  title={A detailed study of {G}aussian Boson Sampling},
   author={Kruse, Regina and Hamilton, Craig S and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
   journal={arXiv preprint arXiv:1801.07488},
   year={2018}
@@ -63,9 +63,9 @@
 
 
 @article{quesada2019,
-  title={Franck-Condon factors by counting perfect matchings of graphs with loops},
+  title={Franck-{C}ondon factors by counting perfect matchings of graphs with loops},
   author={Quesada, N},
-  journal={The Journal of chemical physics},
+  journal={The Journal of Chemical Physics},
   volume={150},
   number={16},
   pages={164113},
@@ -75,7 +75,7 @@
 
 
 @article{quesada2019a,
-  title={Simulating realistic non-Gaussian state preparation},
+  title={Simulating realistic non-{G}aussian state preparation},
   author={Quesada, N and Helt, LG and Izaac, J and Arrazola, JM and Shahrokhshahi, R and Myers, CR and Sabapathy, KK},
   journal={arXiv preprint arXiv:1905.07011},
   year={2019}
@@ -120,7 +120,7 @@
 
 
 @article{dodonov1994,
-  title={Multidimensional Hermite polynomials and photon distribution for polymode mixed light},
+  title={Multidimensional {H}ermite polynomials and photon distribution for polymode mixed light},
   author={Dodonov, VV and Man’ko, OV and Man’ko, VI},
   journal={Physical Review A},
   volume={50},
@@ -158,14 +158,14 @@
 
 @article{quesada2019c,
   author={Quesada, Nicolas and Arrazola, Juan Miguel},
-  title="The Classical Complexity of Gaussian Boson Sampling",
+  title="The Classical Complexity of {G}aussian Boson Sampling",
   year={2019},
   journal={arXiv preprint 1908.08068}
 }
 
 
 @article{gupt2018,
-  title={Classical benchmarking of Gaussian boson sampling on the Titan supercomputer},
+  title={Classical benchmarking of {G}aussian boson sampling on the {T}itan supercomputer},
   author={Gupt, Brajesh and Arrazola, Juan Miguel and Quesada, Nicol{\'a}s and Bromley, Thomas R},
   journal={arXiv preprint arXiv:1810.00900},
   year={2018}
@@ -173,7 +173,7 @@
 
 
 @article{schuld2019,
-  title={A quantum hardware-induced graph kernel based on Gaussian Boson Sampling},
[…]
