### 2019-08-26 577b512
Updates bib file
