+++ paper.bib
@@ -160,7 +160,7 @@
   author={Quesada, Nicolas and Arrazola, Juan Miguel},
   title="The Classical Complexity of {G}aussian Boson Sampling",
   year={2019},
-  journal={arXiv preprint 1908.08068}
+  journal={arXiv preprint arXiv:1908.08068}
 }
 
 
