### 2019-08-28 6559ef1
Update paper.bib
Fixes error in paper.bib
