+++ paper.bib
@@ -197,7 +197,7 @@
 
 
 @article{killoran2019,
-  title={Strawberry fields: {A software platform for photonic quantum computing},
+  title={Strawberry fields: A software platform for photonic quantum computing},
   author={Killoran, Nathan and Izaac, Josh and Quesada, Nicol{\'a}s and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
   journal={Quantum},
   volume={3},
