### 2019-08-29 #1668 @poulson
Oh my -- this is my kind of submission, and even local to Toronto! I apologize for the delay and -- reading through it -- am immediately curious about the interpretations of combining the halfnian generalization of the permanent with the immanantal generalization -- via characters of the symmetric group -- of permanents and determinants. I suppose one would just introduce an arbitrary character of the symmetric group rather than chi(sigma) == 1 in the case of the halfnian.

I will send out invitations to reviewers as soon as I finish reading the paper.
