### 2019-08-29 #1668 @nquesada
Hey @poulson, thanks for agreeing to be editor of our paper; also, very happy to hear you find our paper interesting!
