### 2019-08-29 #1668 @poulson
Hi Max Scheurer (https://github.com/maxscheurer): given your CPPE project (https://github.com/maxscheurer/cppe), which exposes a python interface for an Eigen-based C++ quantum chemistry simulation, it looks like your expertise would be helpful for reviewing this submission.

While the submission is more focused around combinatorial determinant calculations of structured matrices, my guess is that this isn't too much of a stretch for you.

Please let me know if you have the time and interest to review this submission, and would feel confident doing so.
