### 2019-08-29 821d374
Updates bibliography
