+++ paper.bib
@@ -46,7 +46,7 @@
 @inproceedings{bjorklund2012,
   title={Counting perfect matchings as fast as {R}yser},
   author={Bj{\"o}rklund, Andreas},
-  booktitle={Proceedings of the twenty-third annual {ACM-SIAM} symposium on Discrete Algorithms},
+  booktitle={Proceedings of the twenty-third annual ACM-SIAM symposium on Discrete Algorithms},
   pages={914--921},
   year={2012}
 }
@@ -75,11 +75,17 @@
 
 
 @article{quesada2019a,
-  title={Simulating realistic non-{G}aussian state preparation},
-  author={Quesada, N and Helt, LG and Izaac, J and Arrazola, JM and Shahrokhshahi, R and Myers, CR and Sabapathy, KK},
-  journal={arXiv preprint arXiv:1905.07011},
-  year={2019}
+  title = {Simulating realistic non-Gaussian state preparation},
+  author = {Quesada, N. and Helt, L. G. and Izaac, J. and Arrazola, J. M. and Shahrokhshahi, R. and Myers, C. R. and Sabapathy, K. K.},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {2},
+  pages = {022341},
+  numpages = {10},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.022341}
 }
+
 
 
 @article{bjorklund2019,
@@ -197,11 +203,21 @@
 
 
 @article{killoran2019,
-  title={Strawberry fields: {A software platform for photonic quantum computing},
+  title={Strawberry fields: {A} software platform for photonic quantum computing},
   author={Killoran, Nathan and Izaac, Josh and Quesada, Nicol{\'a}s and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
   journal={Quantum},
   volume={3},
   pages={129},
   year={2019},
   doi={10.22331/q-2019-03-11-129}
+}
+
+@article{kok2001,
+  title={Multi-dimensional Hermite polynomials in quantum optics},
+  author={Kok, Pieter and Braunstein, Samuel L},
+  journal={Journal of Physics A: Mathematical and General},
+  volume={34},
+  number={31},
+  pages={6185},
+  year={2001}
 }
