### 2019-08-29 5286de6
Updates paper bib
