+++ docs/references.bib
@@ -482,11 +482,16 @@ edition={3}
 		}
 
 @article{quesada2019simulating,
-  title={Simulating realistic non-Gaussian state preparation},
-    author={Quesada, N and Helt, LG and Izaac, J and Arrazola, JM and Shahrokhshahi, R and Myers, CR and Sabapathy, KK},
-      journal={arXiv preprint arXiv:1905.07011},
-        year={2019}
-	}
+  title = {Simulating realistic non-Gaussian state preparation},
+  author = {Quesada, N. and Helt, L. G. and Izaac, J. and Arrazola, J. M. and Shahrokhshahi, R. and Myers, C. R. and Sabapathy, K. K.},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {2},
+  pages = {022341},
+  numpages = {10},
+  year = {2019}
+}
+
 @article{mizrahi1975generalized,
   title={Generalized hermite polynomials},
   author={Mizrahi, Maurice M},
