### 2019-08-29 bb8d617
Updates in paper.bib
