+++ paper.bib
@@ -46,7 +46,7 @@
 @inproceedings{bjorklund2012,
   title={Counting perfect matchings as fast as {R}yser},
   author={Bj{\"o}rklund, Andreas},
-  booktitle={Proceedings of the twenty-third annual ACM-SIAM symposium on Discrete Algorithms},
+  booktitle={Proceedings of the twenty-third annual {ACM-SIAM} symposium on Discrete Algorithms},
   pages={914--921},
   year={2012}
 }
