### 2019-09-03 #1668 @PhilipVinc
@poulson, should you ask Whedon to start the review issue so that I can fill the checklist?

I have a few minor comments on this library, especially on the documentation, which otherwise seems a very fine piece of software. Normally I would leave those on the Review issue?
