### 2019-09-03 #1668 @katyhuff
Hi Folks (@PhilipVinc @nquesada ), I am stepping in as editor because this content is fairly niche, but Jack (@paulson)and I believe he will be an excellent second reviewer for this work. It was a great idea on Jack's part, and I'm happy to take over handing this one. Now that we have two reviewers and one editor, let's get started on the review.
