### 2019-09-03 #1705 @whedon
[REVIEW]: The Walrus: the fastest calculation of hafnians, Hermite polynomials and Gaussian boson sampling
**Submitting author:** @nquesada (<a href="http://orcid.org/0000-0002-0175-1688">Nicolás Quesada</a>)
**Repository:** <a href="https://github.com/xanaduAI/thewalrus" target ="_blank">https://github.com/xanaduAI/thewalrus</a>
**Version:** v0.10.1
