### 2019-09-10 abba3fb
Adds new references of sfotware/papers using the walrus [ci skip]
