+++ paper.bib
@@ -221,3 +221,18 @@
   pages={6185},
   year={2001}
 }
+
+@article{wu2019speedup,
+  title={Speedup in Classical Simulation of Gaussian Boson Sampling},
+  author={Wu, Bujiao and Cheng, Bin and Zhang, Jialin and Yung, Man-Hong and Sun, Xiaoming},
+  journal={arXiv preprint arXiv:1908.10070},
+  year={2019}
+}
+
+@article{valson2019franck,
+title={Franck-Condon factors via compressive sensing},
+author={{Valson Jacob}, Kevin and {Kaur}, Eneet and {Roga}, Wojciech and
+         {Takeoka}, Masahiro},
+journal={arXiv preprint arXiv:1909.02935},
+year={2019}
+}
