### 2019-09-16 143deeb
Adds integration test
