### 2019-09-16 e249f00
Fixes some minor linting issues
