+++ docs/references.bib
@@ -1,22 +1,19 @@
-% Literature references for the ML Toolbox
-% bibtex file for sphinxcontrib-bibtex
-
 @misc{hafnian,
- author = {Quesada, N. and Gupt, B.},
-    title = {hafnian},
-      year = {2018},
-        publisher = {GitHub},
-	  journal = {GitHub repository},
-	    howpublished = {\url{https://github.com/XanaduAI/hafnian}}
-	   }
+  author = {Quesada, N. and Gupt, B.},
+  title = {hafnian},
+  year = {2018},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/XanaduAI/hafnian}}
+}
 
 @book{termini2006imagination,
-  title={Imagination and Rigor: Essays on Eduardo R. Caianiello's Scientific Heritage
-},
+  title={Imagination and Rigor: Essays on Eduardo R. Caianiello's Scientific Heritage},
   author={Termini, Settimo},
   year={2006},
   publisher={Springer}
 }
+
 @article{bjorklund2018faster,
   title={A faster hafnian formula for complex matrices and its benchmarking on a supercomputer},
   author={Bj{\"o}rklund, Andreas and Gupt, Brajesh and Quesada, Nicol{\'a}s},
@@ -38,6 +35,7 @@
   year={1953},
   publisher={Springer}
 }
+
 @article{banchi2019molecular,
   title={Molecular Docking with Gaussian Boson Sampling},
   author={Banchi, Leonardo and Fingerhuth, Mark and Babej, Tomas and Arrazola, Juan Miguel and others},
@@ -45,89 +43,91 @@
   year={2019}
 }
 
-
 @article{killoran2018strawberry,
   title={Strawberry Fields: A Software Platform for Photonic Quantum Computing},
-    author={Killoran, Nathan and Izaac, Josh and Quesada, Nicol{\'a}s and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
-      journal={arXiv preprint arXiv:1804.03159},
-        year={2018}
-	}
-
+  author={Killoran, Nathan and Izaac, Josh and Quesada, Nicol{\'a}s and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
+  journal={arXiv preprint arXiv:1804.03159},
+  year={2018}
+  }
 
 @book{rempala2007symmetric,
-	title={Symmetric functionals on random matrices and random matchings problems},
-	author={Rempala, Grzegorz and Wesolowski, Jacek},
-	volume={147},
-	year={2007},
-	publisher={Springer Science \& Business Media}
+  title={Symmetric functionals on random matrices and random matchings problems},
+  author={Rempala, Grzegorz and Wesolowski, Jacek},
+  volume={147},
+  year={2007},
+  publisher={Springer Science \& Business Media}
 }
 
 @article{valiant1979complexity,
-	title={The complexity of computing the permanent},
-	author={Valiant, Leslie G},
-	journal={Theoretical computer science},
-	volume={8},
-	number={2},
-	pages={189--201},
-	year={1979},
-	publisher={Elsevier}
+  title={The complexity of computing the permanent},
+  author={Valiant, Leslie G},
+  journal={Theoretical computer science},
+  volume={8},
+  number={2},
+  pages={189--201},
+  year={1979},
+  publisher={Elsevier}
 }
 
 @article{bjorklund2008exact,
-	title={Exact algorithms for exact satisfiability and number of perfect matchings},
-	author={Bj{\"o}rklund, Andreas and Husfeldt, Thore},
-	journal={Algorithmica},
-	volume={52},
-	number={2},
-	pages={226--249},
[…]
