+++ docs/references.bib
@@ -447,7 +447,7 @@ edition={3}
 @article{quesada2019classical,
   author = {Nicol\'as Quesada and Juan Miguel Arrazola},
   title = "The Classical Complexity of Gaussian Boson Sampling",
-  journal={arXiv preprint 1908.08068}
+  journal={arXiv preprint 1908.08068},
   year = {2019}
 }
 
