+++ docs/gallery/gallery.rst
@@ -1,0 +1,39 @@
+.. _gallery:
+
+Nongaussian states gallery
+##########################
+
+
+Here you can find a curated list of circuits and measurements to prepare nongaussian states of interest in quantum optics, information, metrology and metrology. 
+
+
+If you develop a new circuit and measurement scheme to prepare a new nongaussian state add it to the gallery!
+
+
+.. Copy the template below in order to create a link to your notebook, and a thumbnail.
+
+.. _Fock: fock.html
+.. |fock| image:: fock.svg
+   :width: 260px
+   :align: middle
+   :target: fock.html
+
+.. _Cat: cat.html
+.. |cat| image:: cat.svg
+   :width: 260px
+   :align: middle
+   :target: cat.html
+
+.. _Cubic: cubic.html
+.. |cubic| image:: cubic.svg
+   :width: 260px
+   :align: middle
+   :target: cubic.html
+
+
++-------------------------------+-------------------------------+------------------------------+
+| |fock|                        | |cat|                         |   |cubic|                    |
+|                               |                               |                              |
+| Heralded Fock states          | Kitten states                 |   Cubic phase states         |
++-------------------------------+-------------------------------+------------------------------+
+
