### 2019-09-18 99852f8
Fixes typo in cubic phase state and adds photon_added
