### 2019-09-18 c9776ac
Update gallery.rst
