+++ paper.bib
@@ -55,6 +55,17 @@
 
 
 @article{kruse2018,
+  title = {Detailed study of Gaussian boson sampling},
+  author = {Kruse, Regina and Hamilton, Craig S. and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {3},
+  pages = {032326},
+  numpages = {15},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.032326}
+}
+
   title={A detailed study of {G}aussian Boson Sampling},
   author={Kruse, Regina and Hamilton, Craig S and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
   journal={arXiv preprint arXiv:1801.07488},
