+++ docs/references.bib
@@ -259,11 +259,17 @@
 
 
 @article{kruse2018detailed,
-  title={A detailed study of Gaussian Boson Sampling},
-    author={Kruse, Regina and Hamilton, Craig S and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
-      journal={arXiv preprint arXiv:1801.07488},
-        year={2018}
-	}
+  title = {Detailed study of Gaussian boson sampling},
+  author = {Kruse, Regina and Hamilton, Craig S. and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {3},
+  pages = {032326},
+  numpages = {15},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.032326}
+}
+
 @article{2016arXiv160107518B,
 	title={Approximating permanents and hafnians},
 	author={Barvinok, Alexander},
