### 2019-09-18 8180b54
[ci-skip] updates references in paper.bib and references.bib
