+++ paper.bib
@@ -55,10 +55,15 @@
 
 
 @article{kruse2018,
-  title={A detailed study of {G}aussian Boson Sampling},
-  author={Kruse, Regina and Hamilton, Craig S and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
-  journal={arXiv preprint arXiv:1801.07488},
-  year={2018}
+  title = {Detailed study of Gaussian boson sampling},
+  author = {Kruse, Regina and Hamilton, Craig S. and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {3},
+  pages = {032326},
+  numpages = {15},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.032326}
 }
 
 
@@ -220,4 +225,19 @@
   number={31},
   pages={6185},
   year={2001}
+}
+
+@article{wu2019speedup,
+  title={Speedup in Classical Simulation of Gaussian Boson Sampling},
+  author={Wu, Bujiao and Cheng, Bin and Zhang, Jialin and Yung, Man-Hong and Sun, Xiaoming},
+  journal={arXiv preprint arXiv:1908.10070},
+  year={2019}
+}
+
+@article{valson2019franck,
+title={Franck-Condon factors via compressive sensing},
+author={{Valson Jacob}, Kevin and {Kaur}, Eneet and {Roga}, Wojciech and
+         {Takeoka}, Masahiro},
+journal={arXiv preprint arXiv:1909.02935},
+year={2019}
 }
