+++ docs/references.bib
@@ -190,7 +190,7 @@
   author={Ryser, Herbert John},
   number={14},
   year={1963},
-  publisher={Mathematical Association of America; distributed by Wiley [New York}
+  publisher={Mathematical Association of America; distributed by Wiley [New York]}
 }
 
 @book{aaronson2013quantum,
@@ -257,10 +257,15 @@
 }
 
 @article{kruse2018detailed,
-  title={A detailed study of Gaussian Boson Sampling},
-  author={Kruse, Regina and Hamilton, Craig S and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
-  journal={arXiv preprint arXiv:1801.07488},
-  year={2018}
+  title = {Detailed study of Gaussian boson sampling},
+  author = {Kruse, Regina and Hamilton, Craig S. and Sansoni, Linda and Barkhofen, Sonja and Silberhorn, Christine and Jex, Igor},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {3},
+  pages = {032326},
+  numpages = {15},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.032326}
 }
 
 @article{2016arXiv160107518B,
@@ -482,9 +487,14 @@
 
 @article{quesada2019simulating,
   title={Simulating realistic non-Gaussian state preparation},
-  author={Quesada, N and Helt, LG and Izaac, J and Arrazola, JM and Shahrokhshahi, R and Myers, CR and Sabapathy, KK},
-  journal={arXiv preprint arXiv:1905.07011},
-  year={2019}
+  author = {Quesada, N. and Helt, L. G. and Izaac, J. and Arrazola, J. M. and Shahrokhshahi, R. and Myers, C. R. and Sabapathy, K. K.},
+  journal = {Phys. Rev. A},
+  volume = {100},
+  issue = {2},
+  pages = {022341},
+  numpages = {10},
+  year = {2019},
+  doi = {10.1103/PhysRevA.100.022341}
 }
 
 @article{mizrahi1975generalized,
