### 2019-09-18 44f2f0a
Renames cat.* to kitten.* and updates gallery.rst
