### 2019-09-18 e5386f7
Adds cat states
