### 2019-09-22 #1705 @poulson
I am back from overseas travel and am now working on the review. I was able to run the python tests without a hitch, but I filed issue https://github.com/XanaduAI/thewalrus/issues/56 to suggest fixing the hardcoded eigen3 path in `tests/Makefile`.
