### 2019-09-22 XanaduAI/thewalrus#57 @poulson
Running `make` within `examples/` fails due to missing `example-f90` target
It appears that the line
```
all: example-f90 example-cpp
```
in `examples/Makefile` should be modified to remove the `example-f90` target, as there does not seem to be any f90 wrapper for example.o.

This extraneous target leads to an error when running `make` in `examples/` of the form:
```
(base) poulson@bartelby:~/Source/thewalrus/examples$ make
make: *** No rule to make target 'example-f90', needed by 'all'.  Stop.
```
