### 2019-09-22 #1705 @poulson
Submitted https://github.com/XanaduAI/thewalrus/issues/57 regarding the default `Makefile` rule in `examples/` leading to an error.
