### 2019-09-22 XanaduAI/thewalrus#57 @poulson
After manually removing the `example-f90` target, there appear to be two more issues needing to be manually addressed so that the examples will build:

1. The path to eigen3 is hardcoded to `/usr/include/eigen3` as in https://github.com/XanaduAI/thewalrus/issues/56 and should perhaps be `${HOME}/.local/eigen3` or the environment variable `EIGEN_INCLUDE_DIR`.

2. There is no definition of the target `libwalrus.hpp` and removing it from the rule `example.o: example.cpp libwalrus.hpp` seems to fix an error.
