### 2019-09-22 #1705 @poulson
I am very impressed by the library of algorithms for computing halfnians, but I am worried there is insufficient justification for the claim that your library is the "fastest" given the lack of benchmarks relative to pre-existing libraries.

For example, on the subject of computing permanents, you have the wonderful documentation here:
https://the-walrus.readthedocs.io/en/latest/permanent_tutorial.html
but the forecasts for computing the permanent of a 35 x 35 matrix seem to be just under an hour. Whereas the timings at https://codegolf.stackexchange.com/questions/97060/calculate-the-permanent-as-quickly-as-possible?rq=1 seem require just a few minutes.

Also, as noted, both Mathematica and Maple contain library routines for computing permanents. I recognize that these are both proprietary packages -- but I believe the 'fastest' claims demand at least some connection to whatever is closest to state-of-the-art.

As a sidenote, I extended `examples/example.cpp` to compute halfnians up to size 30 x 30 after increasing `nmax` from 10 to 15 and noticed that double-precision numbers overflow starting at `m=15`. Have you considered falling back to higher precision if overflow is detected?
