### 2019-09-22 #1705 @nquesada
Thanks @poulson ! We are working on  fixing the Makefile and also on improving the permanent code. Regarding being "the fastest", I guess we really like the idea of the a Walrus being the fastest, at something ;) . More to the point, we never claimed anything about permanents, they are in the library to for historical reasons, but they are not mentioned in the scope of the library and appear only in the paper submission to provide some historical context. We have done hafnian bench marking (cf. the JEA paper in the bibliography) and also looked at the codegolf implementations.
