### 2019-09-22 #1705 @poulson
Thank you for the fast response, @nquesada! I see what you mean about the claim being specific to hafnians, but you also point out that haf([0, W; W^T, 0]) = perm(W), so it is possible to at least make a qualitative connection between hafnian runtimes and best-in-class permanent runtimes.

But you are right that the codegolf timings at https://codegolf.stackexchange.com/questions/157049/calculate-the-hafnian-as-quickly-as-possible relative to https://the-walrus.readthedocs.io/en/latest/hafnian_tutorial.html and Fig. 5 of https://arxiv.org/pdf/1805.12498.pdf appear more relevant.

It seemed possible that I was misinterpreting the much faster claims of the codegolf, so I downloaded a copy of the C++ source from the "miles" submission and modified the datatype (`TYPE`) from `int` to `double`, and disabled multithreading by making `S` equal to 0. I then generated a random 52 x 52 matrix by forming a 52 x 52 matrix with entries independently drawn from the symmetric normal distribution, added it onto its transpose, then clipped entries to {-1, 0, 1} based upon mapping (-infinity, -1] to -1, (-1, 1) to 0, and [1, infinity) to 1. Running with a single core on my laptop took 67 seconds with Miles's code compiled with -O3.

As an aside, I originally tried for a 104 x 104 matrix, but Miles's code led to a segfault.

I then modified the `examples/example.cpp` driver in `thewalrus` by including `#include <chrono>` then appending the following into the `main` function, just before `return 0;`:
```
    {
      std::vector<double>  mat{
        // The 52^2 {-1, 0, 1} entries of the 52 x 52 matrix go here.
      };
      std::cout << "Starting 52 x 52 hafnian calculation." << std::endl;
      auto start = std::chrono::steady_clock::now();
      double hafval = libwalrus::loop_hafnian(mat);
      auto end = std::chrono::steady_clock::now();
      std::chrono::duration<double> elapsed_seconds = end - start;

      // print out the result
      std::cout << hafval << "(" << elapsed_seconds.count() << " seconds.)"
                << std::endl;
    }
```
I recompiled with `g++ example.cpp -std=c++11 -O3 -Wall -I/usr/include -I../include -I/home/poulson/.local/eigen3 -fopenmp -march=native -c` and the code has been running for quite some time (more than 10 minutes). I can edit in the time when it completes.

Is it possible that the code from miles in the codegolf is incorrect? I am happy to share the random 52 x 52 {-1, 0, 1} matrix I generated and the results from his submission.
