### 2019-09-22 #1705 @poulson
It has been another 21 minutes and thewalrus is still running. I also read a bit more and Miles's code is an implementation of Algorithm 2 of https://arxiv.org/pdf/1107.4466.pdf, which you cite as the "Recursive Algorithm" from [5] in https://the-walrus.readthedocs.io/en/latest/algorithms.html.

However, I see that this algorithm is only defined for non-loop hafnians, and I am comparing it against the loop hafnian. I will redefine a random symmetric 52 x 52 matrix with zero diagonal and retest with Miles's implementation of the recursive algorithm against your `thewalrus::hafnian`, rather than `thewalrus::loop_hafnian`.

Please excuse the confusion!
