### 2019-09-22 #1705 @nquesada
Hi @poulson . The code from @eklotek (aka "miles") is correct. The algorithm he implemented, which comes from a paper by A. Bjorklund, scales like O(n^5 2^{n/2})  where n is the size of the matrix. The code from https://arxiv.org/pdf/1805.12498.pdf scales like O(n^3 2^{n/2}) which is asymptotically faster, but in real life (not n \to \infty) the algorithm implemented by miles and derived by A. Bjorklund in 2012 is faster.
This algorithm, the one derived by A. Bjorklund in 2012 and referred to as the recursive algorithm in the documentation of the walrus is the default option for calculating hafnians. That being said you should not compare against the code golf one because, at least when called from python, all the calculations are done in quad precision. As a matter of fact the the algorithm used by the walrus is a quad precision openMPd version of  the Bjorklund 2012 algorithm implemented by miles. This is all acknowledged in the source code https://github.com/XanaduAI/thewalrus/blob/master/include/recursive_hafnian.hpp

Hope this clarifies the confusion!
