### 2019-09-22 #1705 @nquesada
Just saw your latest message. The problem with loop hafnians is that as far as I know, and I asked A. Bjorklund about this, there is no generalization of the "recursive" algorithm to loop hafnians.
