### 2019-09-22 #1705 @poulson
I looked through the linked source code and you seem to mean `long double` when you say "quad precision", but, according to the wikipedia entry on `long double`: "With gcc on Linux, 80-bit extended precision is the default". This has been my experience in practice, and it seems to only be older chips that supported hardware 128-bit arithmetic which had compilers mapping `long double` to 128-bit IEEE.

I recompiled my modification of Miles's source code to use `TYPE` as `long double` and saw the runtime for the 52 x 52 matrix increase from ~1 minute 20 seconds to ~9 minutes 30 seconds, with the same answer produced in both cases.

But, to be fair, I am calling `example/example.cpp` with an `std::vector<double>` input and believe it is doing its work using `double`, not `long double`.

I have been testing the exact same matrix against `thewalrus::hafnian` (*not* `thewalrus::loop_hafnian`) and it seems to be taking much more time. I now realize that this is due to what you mentioned above, that the asymptotically faster Bjorklund 2012 algorithm is used by `thewalrus::hafnian`, and one must explicitly call `thewalrus::hafnian_recursive`.

When I modified the example driver to call `thewalrus::hafnian_recursive`, it took about 9 minutes, though this is on top of the `loop_hafnian` and `hafnian` routines still running, so take this with a grain of salt. Either way, this is noticeably slower than the Miles code run with `TYPE` equal to `double` (about a minute and 20 seconds).

One simple optimization it seems you could make -- please correct me if I am wrong -- is to change the signature of
https://github.com/XanaduAI/thewalrus/blob/master/include/recursive_hafnian.hpp#L44,
```
template <typename T>
inline T recursive_chunk(std::vector<T> b, int s, int w, std::vector<T> g, int n)
```
to pass `b` and `g` as const references rather than by value since you don't seem to modify either in the routine.
