### 2019-09-22 #1705 @poulson
I think it would be fair to say that it is the responsibility of the unqualified interfaces (i.e., `hafnian` and `loop_hafnian`) to route to the best algorithms based upon the matrix dimensions. There is such a large difference in runtime that such a simple fix seems worthwhile.
