### 2019-09-23 #1705 @poulson
I believe the rest of the time difference -- from @eklotek's 80 seconds up to the "const reference" version of `hafnian_recursive` -- will largely be explained by you reordering the loops in the main computational kernel, which were originally:

```
        for (int u = 0; u < n; u++) {
            TYPE *d = e+u+1,
                  p = g[u], *x = b+(T(s)-1)*m;
            for (int v = 0; v < n-u; v++)
                d[v] += p*x[v];
        }

        for (int j = 1; j < s-2; j++)
            for (int k = 0; k < j; k++)
                for (int u = 0; u < n; u++) {
                    TYPE *d = c+(T(j)+k)*m+u+1,
                          p = b[(T(s-2)+j)*m+u], *x = b+(T(s-1)+k)*m,
                          q = b[(T(s-2)+k)*m+u], *y = b+(T(s-1)+j)*m;
                    for (int v = 0; v < n-u; v++)
                        d[v] += p*x[v] + q*y[v];
                }
```

so that the majority of array accesses are no longer unit stride, leading to large performance degradations due to decreased cache reuse. Said code segment seems to have been translated to

```
    for (u = 0; u < n; u++) {
        for (v = 0; v < n - u; v++) {
            e[u + v + 1] += g[u] * b[v];

            for (j = 1; j < s - 2; j++) {
                for (k = 0; k < j; k++) {
                    c[(n + 1) * (j * (j - 1) / 2 + k) + u + v + 1] +=
                        b[(n + 1) * ((j + 1) * (j + 2) / 2) + u]
                        * b[(n + 1) * ((k + 1) * (k + 2) / 2 + 1) + v]
                        + b[(n + 1) * (k + 1) * (k + 2) / 2 + u]
                        * b[(n + 1) * ((j + 1) * (j + 2) / 2 + 1) + v];
                }
            }
        }
    }
```

with the 'u' loop now on the outside, and the 'k' loop on the interior. Notice that the updates to `c` are now with stride `n+1` instead of stride `1`, and some of the reads of `b` are also now of highly non-uniform stride.

If you preserve the original data access patterns, I suspect you will see a substantial performance improvement.
