### 2019-09-23 #1705 @eklotek
I'm glad to see that code is still of some interest. I'll try to address a few of the concerns. This is all from memory so some parts may be a bit fudged.

The segfault you experienced is most likely due to how I used stack space to store the submatrices that were visited recursively. Switching to malloc there shouldn't encounter that problem, but will be a bit slower with the 2^n times you will be calling malloc/delete on small matrices at the leaves.

The large performance speedup that algorithm 2 demonstrates is from reusing the previously "squeezed" submatrices instead of only applying the squeeze at the end (leaves of the binary tree). Since the values of n are small (even n = 50 for a 100x100 matrix would take over a month on a standard desktop) most speedups aren't realized due to large constant overhead from the setup/initialization/convergence. Using FFT/Karatsuba (in my attempts) doesn't improve convolution speed over the simple discrete pattern (which is also an easy target for vectorization).

The convolution pattern was used after noticing that vectorization wasn't happening good since the data visited had a non-simple access pattern from the four nested loops. Their version hoists the 2 innermost loops out to the top, which affects the matrix since it is no longer being accessed with stride 1. Transposing the matrix could be beneficial.

Thanks for adopting the code even though I never made a proper commit, and thanks for investigating its runtime in more depth. It was originally only intended for integer use, but the TYPE define was for experimenting with floats for FMA vectorization.
