### 2019-09-23 73fbc1a
fixes #57 and #56
