### 2019-09-24 d34e848
Updates bibliography inside quantum.py [ci-skip]
