### 2019-09-24 768e2bd
[ci-skip] removes bullet points inside the tip boxes
