### 2019-09-26 0b265f5
solve issues with pylint
