### 2019-09-27 #1705 @nquesada
Hey @poulson : I looked into the problem of having "overflows" in `examples/example.cpp` when you  use all one matrices of size ~ 30. It turns out it is not an overflow, it is a problem with diagonalizing the very special matrices you get when you pass an all ones matrix; in this case the matrices that pow_trace hast to diagonalize are also all ones matrices and for some reason eigen does not get the right answer. We observed the same problem back in the day when we called LAPACKE from C. To test this I created a matrix with all ones outside the diagonal on some other values in the diagonal. Since the diagonal does not matter you should get the same result (n-1)!! (for an n x n matrix) and indeed in this case you do get the right result. Here is the code:

```
#include <iostream>
#include <complex>
#include <vector>
#include <libwalrus.hpp>
#include <stdlib.h>
#include <time.h>

int main() {

  int m =14;
  // create a 2m*2m all ones matrix
  int n = 2 * m;
  std::vector<std::complex<double>> mat(n * n, 1.0);
  std::vector<std::complex<double>> vec1(n, 0.0);
  /*
  for(int i=0;i<n;i++){
    vec1[i] = 1.0/((double) i);
  }
  for(int i=0;i<n;i++){
    for(int j=0;i<n;i++){
      mat[i,j] = vec1[i]*vec1[j];
    }
  }
  */
  for(int i=0;i<n;i++){
    mat[n*i+i]=1.0/(1.0+i); # set to 0.0 to recover the bug!
  }

  // calculate the hafnian
  std::complex<double> hafval = libwalrus::hafnian(mat);
  // print out the result
  std::cout << n << " " << hafval << std::endl;

  return 0;
};
```

There are at least two ways to go around this. One is to change the signature of all the function to pass  success/fail argument that comes from the diagonalization routines. Another one is to  internally change the values in the diagonal of any input matrix so that they are not all equal.
