### 2019-09-28 #1705 @poulson
@nquesada I would be very surprised if the same bug appeared in LAPACK's Aggressive Early Deflation Hessenberg QR as in Eigen's adaptation of EISPACK via JAMA.

It is worth noting that -- somewhat ironically -- Eigen's generic eigensolver is grossly suboptimal for even moderate sized matrices. But maybe this is irrelevant for you given the small matrices.

But, at this point, my main concern is that your library is currently not justified in its claim to be the "fastest" calculation of "hafnians, Hermite polynomials and Gaussian boson sampling", nor the GitHub repo's claim as being "The fastest exact library for hafnians, torontonians, and permanents for real and complex matrices."

I recommend modifying the recursive hafnian formulation to use unit stride access in the manner originally used by @eklotek (or using its transpose) and making this algorithm the default when one calls `thewalrus::hafnian`, and modifying the GitHub repo's title to only claim the fastest hafnian computation.
