### 2019-10-01 c0b3360
Updates README.md and paper.md
