+++ thewalrus/_hafnian.py
@@ -1,4 +1,3 @@
-# Copyright 2019 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
