### 2019-10-01 b4e4c2b
Updates README, paper.md and docs/index.rst
