### 2019-10-02 4c2643e
Moves definition of datatype qp to stdafx.h
