+++ include/permanent.hpp
@@ -21,12 +21,7 @@ typedef unsigned long long int ullint;
 typedef long long int llint;
 //typedef long double qp;
 
-#if (defined(__GNUC__) || defined(__GNUG__)) && !(defined(__clang__) || defined(__INTEL_COMPILER))
-typedef __float128 qp;
-//#include <quadmath.h>
-#else
-typedef long double qp;
-#endif
+
 
 /**
  * Gray code generator.
