### 2019-10-02 522d87d
Adds test test_number_moments_multimode_thermal that catches bug in generate_thermal_samples
