+++ thewalrus/tests/test_csamples.py
@@ -13,7 +13,10 @@
 # limitations under the License.
 r"""Tests for the classical sampling functions"""
 
+import pytest
+
 import numpy as np
+
 from thewalrus.csamples import (
     rescale_adjacency_matrix_thermal,
     generate_thermal_samples,
@@ -22,12 +25,12 @@ from thewalrus.csamples import (
 rel_tol = 10
 
 def generate_positive_definite_matrix(n):
-    r"""Generates a positive definite matrix of size n
-	Args:
-		n (int) : Size of the matrix
-	Returns:
-		array: Positive definite matrix
-	"""
+    r"""Generates a real positive definite matrix of size n
+    Args:
+        n (int) : Size of the matrix
+    Returns:
+        array: Positive definite matrix
+    """
     A = np.random.rand(n, n)
     return A.T @ A
 
@@ -65,3 +68,28 @@ def test_dist_thermal():
     rel_freq = freq / n_samples
     expected = (1 / (1 + n_mean)) * (n_mean / (1 + n_mean)) ** (np.arange(len(rel_freq)))
     assert np.allclose(rel_freq, expected, atol=10 / np.sqrt(n_samples))
+
+
+@pytest.mark.parametrize("nmodes", [6])
+def test_number_moments_multimode_thermal(nmodes):
+    r"""Test the correct distribution of the photon numbers means and covariances"""
+    n_samples = 100000
+    n_mean = 3.0
+    A = generate_positive_definite_matrix(nmodes)
+    ls, O = rescale_adjacency_matrix_thermal(A, n_mean)
+    Nmat = O @ np.diag(ls/(1.0-ls)) @ O.T
+    samples = np.array(generate_thermal_samples(ls, O, n_samples))
+
+    nmean_est = samples.mean(axis=0)
+    cov_est = np.cov(samples.T)
+    expected_cov = np.zeros_like(cov_est)
+    for i in range(nmodes):
+        for j in range(i):
+            expected_cov[i,j] = Nmat[i,i]*Nmat[j,j]+Nmat[i,j]**2
+            expected_cov[j,i] = expected_cov[i,j]
+        expected_cov[i,i] = 2*Nmat[i,i]**2+Nmat[i,i]
+    ## This moments are obtained using Wick's theorem for a multimode
+    ## thermal states
+    expected_cov = expected_cov - np.outer(np.diag(Nmat), np.diag(Nmat))
+    # To construct the covariance matrix we need to subtract the mean
+    assert np.allclose(nmean_est, np.diag(Nmat), rtol=10/np.sqrt(n_samples))
