+++ thewalrus/tests/test_csamples.py
@@ -17,12 +17,10 @@ import pytest
 
 import numpy as np
 
-from thewalrus.csamples import (
-    rescale_adjacency_matrix_thermal,
-    generate_thermal_samples,
-)
+from thewalrus.csamples import rescale_adjacency_matrix_thermal, generate_thermal_samples
 
 rel_tol = 10
+
 
 def generate_positive_definite_matrix(n):
     r"""Generates a real positive definite matrix of size n
@@ -34,6 +32,7 @@ def generate_positive_definite_matrix(n):
     A = np.random.rand(n, n)
     return A.T @ A
 
+
 def test_rescaling_thermal():
     r"""Test that the rescaled eigenvalues of the matrix give the correct mean photon number for thermal rescaling"""
 
@@ -42,6 +41,7 @@ def test_rescaling_thermal():
     n_mean = 1.0
     ls, _ = rescale_adjacency_matrix_thermal(A, n_mean)
     assert np.allclose(n_mean, np.sum(ls / (1 - ls)))
+
 
 def test_mean_thermal():
     r"""Test that the thermal samples have the correct mean photon number"""
@@ -54,6 +54,7 @@ def test_mean_thermal():
     tot_photon_sample = np.sum(samples, axis=1)
     n_mean_calc = tot_photon_sample.mean()
     assert np.allclose(n_mean_calc, n_mean, rtol=10 / np.sqrt(n_samples))
+
 
 def test_dist_thermal():
     r"""Test that the thermal sampling for a single mode produces the correct photon number distribution"""
@@ -70,14 +71,15 @@ def test_dist_thermal():
     assert np.allclose(rel_freq, expected, atol=10 / np.sqrt(n_samples))
 
 
-@pytest.mark.parametrize("nmodes", [6])
+@pytest.mark.parametrize("nmodes", [3, 4, 5, 6])
 def test_number_moments_multimode_thermal(nmodes):
-    r"""Test the correct distribution of the photon numbers means and covariances"""
+    r"""Test the correct values of the photon number means and covariances"""
+    assert nmodes > 1
     n_samples = 100000
     n_mean = 3.0
     A = generate_positive_definite_matrix(nmodes)
     ls, O = rescale_adjacency_matrix_thermal(A, n_mean)
-    Nmat = O @ np.diag(ls/(1.0-ls)) @ O.T
+    Nmat = O @ np.diag(ls / (1.0 - ls)) @ O.T
     samples = np.array(generate_thermal_samples(ls, O, n_samples))
 
     nmean_est = samples.mean(axis=0)
@@ -85,11 +87,11 @@ def test_number_moments_multimode_thermal(nmodes):
     expected_cov = np.zeros_like(cov_est)
     for i in range(nmodes):
         for j in range(i):
-            expected_cov[i,j] = Nmat[i,i]*Nmat[j,j]+Nmat[i,j]**2
-            expected_cov[j,i] = expected_cov[i,j]
-        expected_cov[i,i] = 2*Nmat[i,i]**2+Nmat[i,i]
-    ## This moments are obtained using Wick's theorem for a multimode
-    ## thermal states
+            expected_cov[i, j] = Nmat[i, i] * Nmat[j, j] + Nmat[i, j] ** 2
+            expected_cov[j, i] = expected_cov[i, j]
+        expected_cov[i, i] = 2 * (Nmat[i, i] ** 2) + Nmat[i, i]
+    ## These moments are obtained using Wick's theorem for a multimode thermal state
     expected_cov = expected_cov - np.outer(np.diag(Nmat), np.diag(Nmat))
-    # To construct the covariance matrix we need to subtract the mean
[…]
