### 2019-10-03 cde51c6
Update test_csamples.py
