### 2019-10-03 93e216b
Updates CHANGELOG.md, version.hpp and _version.py for version 0.8.0
