### 2019-10-03 fbb6f4c
Update .github/CHANGELOG.md
Co-Authored-By: Josh Izaac <josh146@gmail.com>
