### 2019-10-03 4191ca3
Update .github/CHANGELOG.md
Co-Authored-By: Josh Izaac <josh146@gmail.com>
