### 2019-10-03 8fb2411
Updates nongaussian to non-Gaussian
