+++ include/version.hpp
@@ -21,7 +21,7 @@
 #define HAFNIAN_VERSION_MAJOR 0
 
 /// The minor version number
-#define HAFNIAN_VERSION_MINOR 8
+#define HAFNIAN_VERSION_MINOR 9
 
 /// The patch number
 #define HAFNIAN_VERSION_PATCH 0
@@ -30,4 +30,4 @@
 #define HAFNIAN_VERSION_CODE (HAFNIAN_VERSION_MAJOR * 10000 + HAFNIAN_VERSION_MINOR * 100 + HAFNIAN_VERSION_PATCH)
 
 /// Version number as string
-#define HAFNIAN_VERSION_STRING "0.8.0"
+#define HAFNIAN_VERSION_STRING "0.9.0"
