+++ .github/CHANGELOG.md
@@ -1,3 +1,17 @@
+# Version 0.9.0-dev
+
+### New features
+
+### Improvements
+
+### Bug fixes
+
+### Contributors
+
+This release contains contributions from (in alphabetical order):
+
+---
+
 # Version 0.8.0
 
 ### New features
