+++ paper.md
@@ -96,8 +96,8 @@ problem for classical computers, the complexity of the algorithm we provide for 
 complexity of the hafnian, still exponential in the size of the number of particles generated.
 
 Our package has already been used in several research efforts to understand how to generate resource
-states for universal quantum computing [@quesada2019a; @valson2019franck], study the dynamics of vibrational quanta in
-molecules [@quesada2019], and develop the applications of GBS to molecular docking [@banchi2019],
+states for universal quantum computing [@quesada2019a], study the dynamics of vibrational quanta in
+molecules [@quesada2019; @valson2019franck], and develop the applications of GBS to molecular docking [@banchi2019],
 graph theory [@schuld2019], and point processes [@jahangiri2019]. More importantly, it has been
 useful in delineating when quantum computation can be simulated by classical computing resources and
 when it cannot [@gupt2018; @quesada2019c; @wu2019speedup; @killoran2019].
