+++ paper.bib
@@ -48,7 +48,8 @@
   author={Bj{\"o}rklund, Andreas},
   booktitle={Proceedings of the twenty-third annual ACM-SIAM symposium on Discrete Algorithms},
   pages={914--921},
-  year={2012}
+  year={2012},
+  doi = {10.1137/1.9781611973099.73}
 }
 
 
@@ -224,7 +225,8 @@
   volume={34},
   number={31},
   pages={6185},
-  year={2001}
+  year={2001},
+  doi = {10.1088/0305-4470/34/31/312}
 }
 
 @article{wu2019speedup,
