### 2019-10-10 #1705 @nquesada
The missing DOIs have been added in https://github.com/XanaduAI/thewalrus/pull/71
