### 2019-10-10 #1705 @poulson
I am happy to recommend the paper for publication at this point and hope that @PhilipVinc can provide their feedback soon.
