### 2019-10-11 #1705 @katyhuff
@nquesada Thank you for engaging actively in this process. Earlier this week I contacted @PhilipVinc via email in case notifications were not reaching them. Hopefully that review is forthcoming! I will keep you updated.
