### 2019-10-15 6546850
Update gallery.rst
