### 2019-10-15 52e6d32
Update gallery.rst
