### 2019-10-16 93301eb
Update gallery.rst
