### 2019-10-17 d4d6760
Update gallery.rst
