### 2019-10-17 bd7ea7b
Update gallery.rst
