### 2019-10-17 455d749
Update gallery.rst
