### 2019-10-21 #1705 @katyhuff
Dear Xuemei Gu (@XuemeiGu),
I'm emailing to request your expertise in quantum computational physics. Specifically, I'd like to ask for your review of a submission to the [Journal of Open Source Software](http://joss.theoj.org/). I'm one of the editors for JOSS and am in need of a reviewer to evaluate The Walrus: A library for the calculation of hafnians, Hermite polynomials and Gaussian boson sampling. Usually, I would request your review earlier in the process, but this time, we're halfway through the review and we need a replacement reviewer who would be able to look at the software in the next week or two. It is a specialized utility and, because of your quantum computing expertise, I think you would be a phenomenal reviewer. Here is some information about the submission:

- Review Link: https://github.com/openjournals/joss-reviews/issues/1705
- Reviewer Need: This submission is in need of expertise in computational quantum physics as well as familiarity with the systems and languages involved (primarily python).
- Title: The Walrus: the fastest calculation of hafnians, Hermite polynomials and Gaussian boson sampling
- Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01705/joss.01705/10.21105.joss.01705.pdf
- Submitting author: @nquesada (Nicolas Quesada)
- Repository: https://github.com/XanaduAI/thewalrus
- Version: 0.7.0

Here is what you would be agreeing to. First of all, JOSS is not a regular journal. Given your involvement in open-source scientific software development, I imagine you may have heard of it already, but if you haven't, perhaps you'll like the concept: it's an open, developer-friendly journal for open-source research software. The review process is focused on best practices, and emphasizes reviewing the code, rather than reviewing what is written about the code. If the authors have done a good job developing their software, the process should be extremely rapid. [Much more information is available in our reviewer documentation.](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

What do you think?

Would you be willing to review for this paper? I would be delighted for your help! Please let me know in the next few days whether you are available.

Best,
Katy Huff
