### 2019-10-24 3d68a8a
Adds quick_guide
