+++ docs/quick_guide.rst
@@ -1,0 +1,26 @@
+.. _quick_guide:
+
+Quick guide
+###########
+
+This section provides a quick guide to find which function does what in The Walrus.
+
+
+================================================================================ =============================
+**What you want**                                                                **Corresponding function**
+-------------------------------------------------------------------------------- -----------------------------
+Numerical hafnian                                                                :func:`thewalrus.hafnian()`
+Symbolic hafnian                                                                 :func:`thewalrus.reference.hafnian()`
+Hafnian of a matrix with repeated rows and columns                               :func:`thewalrus.hafnian_repeated()`
+Hafnian of all possible reductions of a given matrix                             :func:`thewalrus.hafnian_batched()`
+Hafnian samples of a Gaussian state                                              :func:`thewalrus.samples.hafnian_sample_state()`
+Torontonian samples of a Gaussian state                                          :func:`thewalrus.samples.torontonian_sample_state()`
+Hafnian samples of a graph                                                       :func:`thewalrus.samples.hafnian_sample_graph()`
+Torontonian samples of a graph                                                   :func:`thewalrus.samples.torontonian_sample_graph()`
+All Probability amplitudes of a pure Gaussian state                              :func:`thewalrus.quantum.state_vector()`
+All matrix elements of a general Gaussian state                                  :func:`thewalrus.quantum.density_matrix()`
+A particular probability amplitude of pure Gaussian state                        :func:`thewalrus.quantum.pure_state_amplitude()`
+A particular matrix element of a general Gaussian state                          :func:`thewalrus.quantum.density_matrix_element()`
+================================================================================ =============================
+
+Note that all the hafnian functions listed above generalize to loop hafnians.
