### 2019-10-24 1464722
Update quick_guide.rst
