### 2019-10-27 aa0c9d1
Updates text of Approximate Algorithm [ci-skip]
[ci-skip]
