### 2019-10-29 #1705 @poulson
@katyhuff FWIW, I don't think any quantum computing knowledge is required to review this software; I believe anyone on the mathematical side of numerical linear algebra would be a suitable reviewer; especially anyone with experience computing permanents or with a bit of group theory knowledge (to understand the generalization from determinants to immanents using characters of the symmetric group).
