### 2019-10-30 #1705 @katyhuff
@amitkumarj441 Excellent news. Thank you for agreeing to review. A JOSS review primarily involves checking submissions against a checklist of essential software features and details in the submitted paper and code. Please see the checklist at the top of this issue for your to-do list regarding the review.  More information about the review process can be found in our [reviewer documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).fill out your information at this signup form: https://joss.theoj.org/reviewer-signup.html . If not, please just let me know in this thread.
