### 2019-10-30 #1705 @amitkumarj441
> @amitkumarj441 Excellent news. Thank you for agreeing to review. A JOSS review primarily involves checking submissions against a checklist of essential software features and details in the submitted paper and code. Please see the checklist at the top of this issue for your to-do list regarding the review. More information about the review process can be found in our [reviewer documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Please also, if you have not already, fill out your information at this signup form so that we can put your on our reviewer list, officially: https://joss.theoj.org/reviewer-signup.html .

Thanks for your prompt response.
I have submitted the reviewer form though. But for the checklist, I need [review invitation](https://github.com/openjournals/joss-reviews/invitations) to move forward.

Best Regards,
Amit Kumar Jaiswal
