### 2019-10-30 #1705 @katyhuff
@amitkumarj441 Your review invitation will be approved very soon by one of the @openjournals/joss-eics  . The approval of new reviewers is one of the few parts of the process which is not yet automated.
