### 2019-10-31 #1705 @arfon
@amitkumarj441 - you should have an invite to accept at https://github.com/openjournals/joss-reviews/invitations now.
