### 2019-11-01 5ec192a
Adds tests for rotation, beam_splitter and expand_vector in symplectic
