+++ thewalrus/tests/test_symplectic.py
@@ -1,0 +1,618 @@
+# Copyright 2019 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Basic chip tests"""
+# pylint: disable=no-self-use, assignment-from-no-return
+import pytest
+
+import numpy as np
+from scipy.linalg import block_diag
+
+from thewalrus import symplectic
+
+
+# defaults
+TOL = 1e-3
+
+
+@pytest.fixture(scope="session")
+def tol():
+    """Numerical tolerance for equality tests."""
+    return TOL
+
+
+@pytest.fixture(params=[0.5, 1, 2])
+def hbar(request):
+    """The value of hbar to use in tests"""
+    return request.param
+
+# make test deterministic
+np.random.seed(42)
+
+
+class TestVacuum:
+    """Tests for the vacuum_state function"""
+
+    def test_vacuum_state(self, hbar, tol):
+        """Test the vacuum state is correct."""
+        modes = 3
+        means, cov = symplectic.vacuum_state(modes, hbar=hbar)
+        assert np.allclose(means, np.zeros([2 * modes]), atol=tol, rtol=0)
+        assert np.allclose(cov, np.identity(2 * modes) * hbar / 2, atol=tol, rtol=0)
+
+
+class TestSqueezing:
+    """Tests for the squeezing symplectic"""
+
+    def test_symplectic(self, tol):
+        """Test that the squeeze operator is symplectic"""
+        r = 0.543
+        phi = 0.123
+        S = symplectic.squeezing(r, phi)
+
+        # the symplectic matrix
+        O = np.array([[0, 1], [-1, 0]])
+
+        assert np.allclose(S @ O @ S.T, O, atol=tol, rtol=0)
+
+    def test_squeezing(self, tol):
+        """Test the squeezing symplectic transform."""
+        r = 0.543
+        phi = 0.123
+        S = symplectic.squeezing(r, phi)
+        out = S @ S.T
+
+        # apply to an identity covariance matrix
+        rotation = np.array([[np.cos(phi/2), -np.sin(phi/2)], [np.sin(phi/2), np.cos(phi/2)]])
+        expected = rotation @ np.diag(np.exp([-2*r, 2*r])) @ rotation.T
+        assert np.allclose(out, expected, atol=tol, rtol=0)
+
+
+class TestTwoModeSqueezing:
+    """Tests for the TMS symplectic"""
+
+    def test_symplectic(self, tol):
+        """Test that the two mode squeeze operator is symplectic"""
+        r = 0.543
+        phi = 0.123
+        S = symplectic.expand(symplectic.two_mode_squeezing(r, phi), modes=[0, 2], N=4)
+
+        # the symplectic matrix
+        O = np.block([[np.zeros([4, 4]), np.identity(4)], [-np.identity(4), np.zeros([4, 4])]])
+
+        assert np.allclose(S @ O @ S.T, O, atol=tol, rtol=0)
+
+    def test_decompose(self, tol):
[…]
