+++ thewalrus/symplectic.py
@@ -1,0 +1,269 @@
+# Copyright 2019 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+Symplectic
+==========
+
+**Module name:** :mod:`thewalrus.symplectic`
+
+.. currentmodule:: thewalrus.symplectic
+
+Contains some Gaussian operations and auxiliary functions.
+
+Auxiliary functions
+-------------------
+
+.. autosummary::
+    expand
+    reduced_state
+
+Gaussian states
+---------------
+
+.. autosummary::
+    vacuum_state
+
+
+Gates and operations
+--------------------
+
+.. autosummary::
+    two_mode_squeezing
+    interferometer
+    loss
+    mean_photon_number
+    beam_splitter
+    rotation
+
+Code details
+^^^^^^^^^^^^
+"""
+import numpy as np
+
+
+def expand(S, modes, N):
+    r"""Expands a Symplectic matrix S to act on the entire subsystem.
+
+    Args:
+        S (array): a :math:`2M\times 2M` Symplectic matrix
+        modes (Sequence[int]): the list of modes S acts on
+        N (int): full size of the subsystem
+
+    Returns:
+        array: the resulting :math:`2N\times 2N` Symplectic matrix
+    """
+    M = len(S) // 2
+    S2 = np.identity(2 * N)
+    w = np.array(modes)
+
+    S2[w.reshape(-1, 1), w.reshape(1, -1)] = S[:M, :M].copy()  # X
+    S2[(w + N).reshape(-1, 1), (w + N).reshape(1, -1)] = S[M:, M:].copy()  # P
+    S2[w.reshape(-1, 1), (w + N).reshape(1, -1)] = S[:M, M:].copy()  # XP
+    S2[(w + N).reshape(-1, 1), w.reshape(1, -1)] = S[M:, :M].copy()  # PX
+
+    return S2
+
+def expand_vector(alpha, mode, N, hbar=2.0):
+    r = np.zeros(2*N)
+    r[mode] = np.sqrt(2*hbar)*alpha.real
+    r[N+mode] = np.sqrt(2*hbar)*alpha.imag
+    return r
+
+def reduced_state(mu, cov, modes):
+    r""" Returns the vector of means and the covariance matrix of the specified modes.
+
+    Args:
+        modes (int of Sequence[int]): indices of the requested modes
+
+    Returns:
+        tuple (means, cov): means is an array containing the vector of means,
+        and cov is a square array containing the covariance matrix. Both use the :math:`xp`-ordering.
+    """
+    N = len(mu) // 2
+
+    if modes == list(range(N)):
+        # reduced state is full state
+        return mu, cov
+
+    # reduce state down to specified subsystems
+    if isinstance(modes, int):
+        modes = [modes]
+
+    if np.any(np.array(modes) > N):
+        raise ValueError("The specified modes cannot be larger than number of subsystems.")
+
+    ind = np.concatenate([np.array(modes), np.array(modes) + N])
+    rows = ind.reshape(-1, 1)
+    cols = ind.reshape(1, -1)
+
+    return mu.copy()[ind], cov.copy()[rows, cols]
+
+
[…]
