### 2019-11-01 6af5d00
Moves fixtures from test_symplectic to conftest
