### 2019-11-01 57ba9b0
Moves fixtures from test_symplectic to conftest
