### 2019-11-01 17e787f
Fixes linting issues
