+++ thewalrus/tests/conftest.py
@@ -20,24 +20,20 @@ import numpy as np
 # defaults
 TOL = 1e-3
 
-
 @pytest.fixture(scope="session")
 def tol():
     """Numerical tolerance for equality tests."""
     return TOL
-
 
 @pytest.fixture(params=[0.5, 1, 2])
 def hbar(request):
     """The value of hbar to use in tests"""
     return request.param
 
-
 @pytest.fixture(params=[np.complex128, np.float64, np.int64])
 def dtype(request):
     """Fixture that iterates through all numpy types"""
     return request.param
-
 
 @pytest.fixture
 def random_matrix(dtype):
@@ -56,4 +52,3 @@ def random_matrix(dtype):
         return dtype(A)
 
     return _wrapper
-
