### 2019-11-01 e978101
Corrects incorrectly names tests
