### 2019-11-01 be8910e
Updates based on code review
