### 2019-11-01 9db2ede
Update .github/CHANGELOG.md
Co-Authored-By: Josh Izaac <josh146@gmail.com>
