### 2019-11-01 1915665
Removes lines from CHANGELOG
