### 2019-11-01 e2878c9
Updates to development version
