+++ include/version.hpp
@@ -30,4 +30,4 @@
 #define HAFNIAN_VERSION_CODE (HAFNIAN_VERSION_MAJOR * 10000 + HAFNIAN_VERSION_MINOR * 100 + HAFNIAN_VERSION_PATCH)
 
 /// Version number as string
-#define HAFNIAN_VERSION_STRING "0.9.0"
+#define HAFNIAN_VERSION_STRING "0.10.0"
