### 2019-11-04 cd35742
Makes the interface of all the gates uniform. They now all return real symplectic matrices
