### 2019-11-04 aeb4490
Removes commented line
