+++ thewalrus/tests/test_operations.py
@@ -1,0 +1,140 @@
+# Copyright 2019 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# pylint: disable=no-self-use, assignment-from-no-return
+import pytest
+
+import numpy as np
+
+from thewalrus import operations
+from thewalrus.symplectic import rotation, squeezing, interferometer
+
+# make test deterministic
+np.random.seed(137)
+
+
+@pytest.mark.parametrize("cutoff", [4, 5, 6, 7])
+def test_single_mode_identity(cutoff, tol):
+    nmodes = 1
+    S = np.identity(2 * nmodes)
+    alphas = np.zeros([nmodes])
+    T = operations.fock_tensor(S, alphas, cutoff)
+    expected = np.identity(cutoff)
+    assert np.allclose(T, expected, atol=tol, rtol=0)
+
+
+@pytest.mark.parametrize("cutoff", [7])
+def test_single_mode_rotation(cutoff, tol):
+    nmodes = 1
+    theta = 2 * np.pi * np.random.rand()
+    S = rotation(theta)
+    alphas = np.zeros([nmodes])
+    T = operations.fock_tensor(S, alphas, cutoff)
+    expected = np.diag(np.exp(1j * theta * np.arange(cutoff)))
+    assert np.allclose(T, expected, atol=tol, rtol=0)
+
+
+@pytest.mark.parametrize("cutoff", [5])
+def test_single_mode_displacement(cutoff, tol):
+    nmodes = 1
+    alphas = (0.3 + 0.5 * 1j) * np.ones([nmodes])
+    S = np.identity(2 * nmodes)
+    # This data is obtained by using qutip
+    # np.array(displace(40,alpha).data.todense())[0:10,0:10]
+    data = np.load("displacement.npy")
+    if cutoff <= len(data):
+        expected = data[0:cutoff, 0:cutoff]
+        T = operations.fock_tensor(S, alphas, cutoff)
+        assert np.allclose(T, expected, atol=tol, rtol=0)
+
+
+@pytest.mark.parametrize("cutoff", [5])
+def test_single_mode_squeezing(cutoff, tol):
+    nmodes = 1
+    r = 1.0
+    S = squeezing(r, 0.0)
+    alphas = np.zeros([nmodes])
+    # This data is obtained by using qutip
+    # np.array(squeeze(40,1.0).data.todense())[0:10,0:10]
+    data = np.load("squeeze.npy")
+    if cutoff <= len(data):
+        expected = data[0:cutoff, 0:cutoff]
+        T = operations.fock_tensor(S, alphas, cutoff)
+        assert np.allclose(T, expected, atol=tol, rtol=0)
+
+
+@pytest.mark.parametrize("cutoff", [5])
+def test_single_mode_displacement_squeezing(cutoff, tol):
+    nmodes = 1
+    r = 1.0
+    S = squeezing(r, 0.0)
+    alphas = (0.5 + 0.4 * 1j) * np.ones([nmodes])
+    # This data is obtained by using qutip
+    # np.array((displace(40,alpha)*squeeze(40,r)).data.todense())[0:10,0:10]
[…]
