### 2019-11-06 b5bf650
Fixes asserts in tests
