### 2019-11-06 7551bea
Adds module docstring in test_operations
