### 2019-11-06 9346b25
Updates docstring of test_symplectic
