+++ thewalrus/tests/test_symplectic.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Basic chip tests"""
+"""Symplectic tests"""
 # pylint: disable=no-self-use, assignment-from-no-return
 import pytest
 
