### 2019-11-07 XanaduAI/thewalrus#84 @amitkumarj441
Fix Recurrence
This statement "the analogous probability **for for** the case of threshold detectors" has word recurrence that needs [fixing](https://github.com/XanaduAI/thewalrus/blob/master/docs/gbs.rst) here: https://the-walrus.readthedocs.io/en/latest/gbs.html#gaussian-states-and-threshold-detection

Also, in [Notation](https://the-walrus.readthedocs.io/en/latest/notation.html#the-reduction-operation-in-terms-of-sets) documentation, I think "(multiset)set" should be either "(multiset) set" or "multiset".
