### 2019-11-07 0373cb8
Fixes typos
Addresses https://github.com/XanaduAI/thewalrus/issues/84
