### 2019-11-07 8efa474
Update notation.rst
Addresses https://github.com/XanaduAI/thewalrus/issues/84
