### 2019-11-07 XanaduAI/thewalrus#84 @nquesada
Thanks @amitkumarj441. These typos are fixed in https://github.com/XanaduAI/thewalrus/pull/85
