### 2019-11-07 #1705 @amitkumarj441
I have been reviewing the [thewalrus](https://github.com/XanaduAI/thewalrus) library and paper, and find it as one of the best package to compute constructs such as hafnians, Hermite polynomials and torontorians for simulating a quantum framework/computer based on a Gaussian boson sampler. The preliminaries for this library and examples are well documented. I have submitted this issue [XanaduAI/thewalrus/#84](https://github.com/XanaduAI/thewalrus/issues/84) to update the documentation with minor fixes.

I'm happy with the library - it installs and tests run. I think the current version of the paper looks good.

I had a quick read of this paper: https://arxiv.org/pdf/1910.04022.pdf and I think this should be referenced in the paper as well in **a matrix function that counts the number of perfect matchings of a graph that has loops**. I might be completely wrong here.

Also, I have calculated the probabilities of an existing GBS experiment using the thewalrus library to examine the zero-mean Gaussian state~($\psi_{g_{0}}$) (as written in the paper - at the start of 3rd paragraph from the last paragraph) if it supports the reported optimal methods.

Considering this equation
`$\sum(\vec{x}_\alpha)$ = $\frac{1}{(2\pi)^N} \braket{\vec{\alpha}|\psi_{g_{0}}} \braket{\psi_{g_{0}}|\vec{\alpha}}$` as a Gaussian function. Then I calculate the K function (a unique way to represent a pure state which is a square root of the Husimi function up to a phase) given as `$K(\vec{x}_\alpha)$ = $\frac{1}{(2\pi)^N} \braket{\vec{\alpha}|\psi_{g_{0}}}` by splitting the Gaussian function ($\sum$) into two conjugate parts. It is simplified in the paper by using complex coordinates.

It would be good to have at least one general example for a pure Gaussian state of N-mode with covariance matrix apart from already given [specific examples](https://the-walrus.readthedocs.io/en/latest/gbs.html#gaussian-states-and-threshold-detection) to support the statement _The relation between GBS and hafnians stems from the fact that the probability of a given experimental outcome is proportional to the hafnian of a matrix constructed from the covariance matrix of the Gaussian state. This observation requires the Gaussian state to have zero mean and the detector to be able to resolve any number of particles_ written in the paper.
