### 2019-11-07 #1705 @nquesada
Hi @amitkumarj441 . There are a number of examples of Gaussian states where we explicitly show the covariance matrix in The Walrus Gallery https://the-walrus.readthedocs.io/en/latest/gallery/gallery.html
