### 2019-11-07 #1705 @nquesada
Also, we just merged the PR correcting the two typos you found in the documentation.
