### 2019-11-07 f28b45c
Hardcodes the expected values into the test file
