+++ thewalrus/tests/test_operations.py
@@ -47,51 +47,69 @@ def test_single_mode_rotation(cutoff, tol):
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("cutoff", [5])
-def test_single_mode_displacement(cutoff, tol):
+def test_single_mode_displacement(tol):
     """Tests the correct construction of the single mode displacement operation"""
     nmodes = 1
+    cutoff = 5
     alphas = (0.3 + 0.5 * 1j) * np.ones([nmodes])
     S = np.identity(2 * nmodes)
     # This data is obtained by using qutip
-    # np.array(displace(40,alpha).data.todense())[0:10,0:10]
-    data = np.load("displacement.npy")
-    if cutoff <= len(data):
-        expected = data[0:cutoff, 0:cutoff]
-        T = operations.fock_tensor(S, alphas, cutoff)
-        assert np.allclose(T, expected, atol=tol, rtol=0)
+    # np.array(displace(40,alpha).data.todense())[0:5,0:5]
+    expected = np.array(
+        [
+            [0.84366482 + 0.00000000e00j, -0.25309944 + 4.21832408e-01j, -0.09544978 - 1.78968334e-01j, 0.06819609 + 3.44424719e-03j, -0.01109048 + 1.65323865e-02j],
+            [0.25309944 + 4.21832408e-01j, 0.55681878 + 0.00000000e00j, -0.29708743 + 4.95145724e-01j, -0.14658716 - 2.74850926e-01j, 0.12479885 + 6.30297236e-03j],
+            [-0.09544978 + 1.78968334e-01j, 0.29708743 + 4.95145724e-01j, 0.31873657 + 0.00000000e00j, -0.29777767 + 4.96296112e-01j, -0.18306015 - 3.43237787e-01j],
+            [-0.06819609 + 3.44424719e-03j, -0.14658716 + 2.74850926e-01j, 0.29777767 + 4.96296112e-01j, 0.12389162 + 1.10385981e-17j, -0.27646677 + 4.60777945e-01j],
+            [-0.01109048 - 1.65323865e-02j, -0.12479885 + 6.30297236e-03j, -0.18306015 + 3.43237787e-01j, 0.27646677 + 4.60777945e-01j, -0.03277289 + 1.88440656e-17j],
+        ]
+    )
+    T = operations.fock_tensor(S, alphas, cutoff)
+    assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("cutoff", [5])
-def test_single_mode_squeezing(cutoff, tol):
+def test_single_mode_squeezing(tol):
     """Tests the correct construction of the single mode squeezing operation"""
     nmodes = 1
     r = 1.0
+    cutoff = 5
     S = squeezing(r, 0.0)
     alphas = np.zeros([nmodes])
     # This data is obtained by using qutip
-    # np.array(squeeze(40,1.0).data.todense())[0:10,0:10]
-    data = np.load("squeeze.npy")
-    if cutoff <= len(data):
-        expected = data[0:cutoff, 0:cutoff]
-        T = operations.fock_tensor(S, alphas, cutoff)
-        assert np.allclose(T, expected, atol=tol, rtol=0)
+    # np.array(squeeze(40,r).data.todense())[0:5,0:5]
+    expected = np.array(
+        [
+            [0.80501818 + 0.0j, 0.0 + 0.0j, 0.43352515 + 0.0j, 0.0 + 0.0j, 0.2859358 + 0.0j],
[…]
