### 2019-11-07 #1705 @amitkumarj441
Thanks @nquesada for your prompt response.

I have seen the gallery but didn't inspect it further.

How about this :arrow_down:?
> I had a quick read of this paper: https://arxiv.org/pdf/1910.04022.pdf and I think this should be referenced in the paper as well in **a matrix function that counts the number of perfect matchings of a graph that has loops**. I might be completely wrong here.
>
