### 2019-11-07 #1705 @nquesada
We have not implemented anything from that paper.
