### 2019-11-07 #1705 @nquesada
...and they idea of perfect matchings including loops was introduced in Bjorklund, Gupt and Quesada, which seems to me like the relevant reference.
