### 2019-11-11 87b3daa
Adds symplectic matrix in symplectic.py and improves performance of n_two_mode_squeezed_vac
