### 2019-11-11 1801533
Adds symplectic matrix in symplectic.py and improves performance of n_two_mode_squeezed_vac
