### 2019-11-11 2ba68d0
Applies black to operations.py
