### 2019-11-11 b6705c4
Fixes missing break line in symplectic
