### 2019-11-11 939e59d
Implements Antal's solution
