### 2019-11-11 8bc2b1d
Removes the now deprecated renormalizer function
