### 2019-11-11 ca0302b
Massive improvements in speed by being clever about accessing the elements of the tensor. Huge thanks to Antal!
