### 2019-11-12 0bb3aae
Update thewalrus/operations.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
