### 2019-11-12 383a744
Update thewalrus/operations.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
