### 2019-11-12 63a8457
Update thewalrus/operations.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
