### 2019-11-12 cfff4e7
Update thewalrus/symplectic.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
