### 2019-11-12 d613b61
Update thewalrus/operations.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
