### 2019-11-12 8c29df6
Update thewalrus/symplectic.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
