### 2019-11-12 4dc8637
Updates operations and test_operations based on code review
