+++ thewalrus/tests/test_operations.py
@@ -14,40 +14,65 @@
 """Operations tests"""
 # pylint: disable=no-self-use, assignment-from-no-return
 import pytest
+from itertools import product
 
 import numpy as np
+from scipy.linalg import qr
 
 from thewalrus import operations
-from thewalrus.symplectic import rotation, squeezing, interferometer
+from thewalrus.symplectic import rotation, squeezing, interferometer, two_mode_squeezing, beam_splitter
 
 # make test deterministic
 np.random.seed(137)
 
 
-@pytest.mark.parametrize("cutoff", [4, 5, 6, 7])
-def test_single_mode_identity(cutoff, tol):
+def random_interferometer(N, real=False):
+    r"""Random unitary matrix representing an interferometer.
+
+    For more details, see :cite:`mezzadri2006`.
+
+    Args:
+        N (int): number of modes
+        real (bool): return a random real orthogonal matrix
+
+    Returns:
+        array: random :math:`N\times N` unitary distributed with the Haar measure
+    """
+    if real:
+        z = np.random.randn(N, N)
+    else:
+        z = (np.random.randn(N, N) + 1j * np.random.randn(N, N)) / np.sqrt(2.0)
+    q, r = qr(z)
+    d = np.diag(r)
+    ph = d / np.abs(d)
+    U = np.multiply(q, ph, q)
+    return U
+
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_identity(r, tol):
     """Tests the correct construction of the single mode identity operation"""
     nmodes = 1
+    cutoff = 7
     S = np.identity(2 * nmodes)
     alphas = np.zeros([nmodes])
-    T = operations.fock_tensor(S, alphas, cutoff)
+    T = operations.fock_tensor(S, alphas, cutoff, r=r)
     expected = np.identity(cutoff)
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
-
-@pytest.mark.parametrize("cutoff", [7])
-def test_single_mode_rotation(cutoff, tol):
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_rotation(r, tol):
     """Tests the correct construction of the single mode rotation operation"""
     nmodes = 1
+    cutoff = 7
     theta = 2 * np.pi * np.random.rand()
     S = rotation(theta)
     alphas = np.zeros([nmodes])
-    T = operations.fock_tensor(S, alphas, cutoff)
+    T = operations.fock_tensor(S, alphas, cutoff, r=r)
     expected = np.diag(np.exp(1j * theta * np.arange(cutoff)))
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
-
-def test_single_mode_displacement(tol):
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_displacement(r, tol):
     """Tests the correct construction of the single mode displacement operation"""
     nmodes = 1
     cutoff = 5
@@ -64,16 +89,16 @@ def test_single_mode_displacement(tol):
             [-0.01109048 - 1.65323865e-02j, -0.12479885 + 6.30297236e-03j, -0.18306015 + 3.43237787e-01j, 0.27646677 + 4.60777945e-01j, -0.03277289 + 1.88440656e-17j],
         ]
     )
-    T = operations.fock_tensor(S, alphas, cutoff)
+    T = operations.fock_tensor(S, alphas, cutoff, r=r)
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
-
[…]
