+++ thewalrus/operations.py
@@ -29,7 +29,6 @@ Auxiliary functions
 -------------------
 
 .. autosummary::
-    n_two_mode_squeezed_vac
     choi_expand
 
 Operations
@@ -44,25 +43,8 @@ from itertools import product
 import numpy as np
 
 from thewalrus import hafnian_batched
-from thewalrus.symplectic import expand
+from thewalrus.symplectic import expand, is_symplectic
 from thewalrus.quantum import Amat
-
-
-def n_two_mode_squeezed_vac(n, r=np.arcsinh(1.0)):
-    r"""
-    Returns the symplectic matrix associated with two mode squeezing operations.
-
-    Args:
-        n (integer): number of modes
-        r (float): squeezing parameter
-    Returns:
-        (array): the symplectic matrix
-    """
-    ch = np.cosh(r) * np.identity(n)
-    sh = np.sinh(r) * np.identity(n)
-    zh = np.zeros([n, n])
-    Snet = np.block([[ch, sh, zh, zh], [sh, ch, zh, zh], [zh, zh, ch, -sh], [zh, zh, -sh, ch]])
-    return Snet
 
 
 def choi_expand(S, r=np.arcsinh(1.0)):
@@ -77,42 +59,85 @@ def choi_expand(S, r=np.arcsinh(1.0)):
     """
 
     (n, m) = S.shape
-    assert n == m
-    assert n % 2 == 0
+    if n != m:
+        raise ValueError("The matrix S is not square")
+    if n % 2 != 0:
+        raise ValueError("The matrix S is not of even size")
+
     nmodes = n // 2
-    return expand(S, list(range(nmodes)), n) @ n_two_mode_squeezed_vac(nmodes, r)
+
+    ch = np.cosh(r) * np.identity(nmodes)
+    sh = np.sinh(r) * np.identity(nmodes)
+    zh = np.zeros([nmodes, nmodes])
+    Schoi = np.block([[ch, sh, zh, zh], [sh, ch, zh, zh], [zh, zh, ch, -sh], [zh, zh, -sh, ch]])
+
+    return expand(S, list(range(nmodes)), n) @ Schoi
 
 
-def fock_tensor(S, alpha, cutoff, r=np.arcsinh(1.0)):
+def fock_tensor(S, alpha, cutoff, r=np.arcsinh(1.0), check_symplectic=True, rtol=1e-05, atol=1e-08):
     r"""
-    Calculates the Fock representation of a Gaussian unitary.
-    and displacements alpha up to cutoff
+    Calculates the Fock representation of a Gaussian unitary parametrized by
+    the symplectic matrix S and the displacements alpha up to cutoff in Fock space.
+    For a complete description of what is being done once the matrix B is obtained
+    see
+
+    * Quesada, N.
+    "Franck-Condon factors by counting perfect matchings of graphs with loops."
+    `Journal of Chemical Physics 150, 164113 (2019) <https://aip.scitation.org/doi/10.1063/1.5086387>`_
+
+
     Args:
         S (array): symplectic matrix
         alpha (array): complex vector of displacements
         cutoff (int): cutoff in Fock space
         r (float): squeezing parameter used for the Choi expansion
+        check_symplectic (boolean): checks whether the input matrix is symplectic
     Return:
         (array): Tensor containing the Fock representation of the Gaussian unitary
     """
+    # Check the matrix is symplectic
+    if check_symplectic:
+        if not is_symplectic(S, rtol=1e-05, atol=1e-08):
+            raise ValueError("The matrix S is not symplectic")
+
+    # And that S and alpha have compatible dimensions
+    l, _ = S.shape
+    if l // 2 != len(alpha):
+        raise ValueError("The matrix S and the vector alpha do not have compatible dimensions")
+
[…]
