### 2019-11-12 7ce689c
Adds missing links for docs to be built
