+++ docs/code/operations.rst
@@ -1,0 +1,2 @@
+.. automodule:: thewalrus.operations
+	:members:
