### 2019-11-12 600282c
Solves linting issues
