### 2019-11-12 4dcd293
Corrects typos in docstrings of tests
