### 2019-11-12 4db3747
removes extra args of fock_tensor
