### 2019-11-12 153691f
Simpler way of getting fock_tensor added
