+++ thewalrus/tests/test_quantum.py
@@ -13,9 +13,15 @@
 # limitations under the License.
 """Tests for The Walrus quantum functions"""
 # pylint: disable=no-self-use,redefined-outer-name
+from itertools import product
+
 import pytest
 
 import numpy as np
+from scipy.linalg import qr
+
+from thewalrus.symplectic import rotation, squeezing, interferometer, two_mode_squeezing, beam_splitter
+
 from thewalrus.quantum import (
     reduced_gaussian,
     Xmat,
@@ -37,11 +43,12 @@ from thewalrus.quantum import (
     is_classical_cov,
     total_photon_num_dist_pure_state,
     gen_single_mode_dist,
+    fock_tensor,
 )
 
 
 # make tests deterministic
-np.random.seed(42)
+np.random.seed(137)
 
 
 def TMS_cov(r, phi):
@@ -689,3 +696,198 @@ def test_total_photon_num_dist_pure_state(cutoff):
     p1 = total_photon_num_dist_pure_state(cov, cutoff=cutoff)
     p2 = gen_single_mode_dist(np.arcsinh(np.sqrt(nmean)), N=n, cutoff=cutoff)
     assert np.allclose(p1, p2)
+
+
+
+
+def random_interferometer(N, real=False):
+    r"""Random unitary matrix representing an interferometer.
+
+    For more details, see :cite:`mezzadri2006`.
+
+    Args:
+        N (int): number of modes
+        real (bool): return a random real orthogonal matrix
+
+    Returns:
+        array: random :math:`N\times N` unitary distributed with the Haar measure
+    """
+    if real:
+        z = np.random.randn(N, N)
+    else:
+        z = (np.random.randn(N, N) + 1j * np.random.randn(N, N)) / np.sqrt(2.0)
+    q, r = qr(z)
+    d = np.diag(r)
+    ph = d / np.abs(d)
+    U = np.multiply(q, ph, q)
+    return U
+
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_identity(r, tol):
+    """Tests the correct construction of the single mode identity operation"""
+    nmodes = 1
+    cutoff = 7
+    S = np.identity(2 * nmodes)
+    alphas = np.zeros([nmodes])
+    T = fock_tensor(S, alphas, cutoff, r=r)
+    expected = np.identity(cutoff)
+    assert np.allclose(T, expected, atol=tol, rtol=0)
+
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_rotation(r, tol):
+    """Tests the correct construction of the single mode rotation operation"""
+    nmodes = 1
+    cutoff = 7
+    theta = 2 * np.pi * np.random.rand()
+    S = rotation(theta)
+    alphas = np.zeros([nmodes])
+    T = fock_tensor(S, alphas, cutoff, r=r)
+    expected = np.diag(np.exp(1j * theta * np.arange(cutoff)))
+    assert np.allclose(T, expected, atol=tol, rtol=0)
+
+@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_displacement(r, tol):
+    """Tests the correct construction of the single mode displacement operation"""
+    nmodes = 1
+    cutoff = 5
+    alphas = (0.3 + 0.5 * 1j) * np.ones([nmodes])
+    S = np.identity(2 * nmodes)
+    # This data is obtained by using qutip
[…]
