### 2019-11-12 7ffc0d4
removes operations
