### 2019-11-12 8dbfe75
removes test_operations
