+++ thewalrus/tests/test_operations.py
@@ -1,221 +1,0 @@
-# Copyright 2019 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Operations tests"""
-# pylint: disable=no-self-use, assignment-from-no-return
-
-from itertools import product
-
-import pytest
-
-import numpy as np
-from scipy.linalg import qr
-
-from thewalrus import operations
-from thewalrus.symplectic import rotation, squeezing, interferometer, two_mode_squeezing, beam_splitter
-
-# make test deterministic
-np.random.seed(137)
-
-
-def random_interferometer(N, real=False):
-    r"""Random unitary matrix representing an interferometer.
-
-    For more details, see :cite:`mezzadri2006`.
-
-    Args:
-        N (int): number of modes
-        real (bool): return a random real orthogonal matrix
-
-    Returns:
-        array: random :math:`N\times N` unitary distributed with the Haar measure
-    """
-    if real:
-        z = np.random.randn(N, N)
-    else:
-        z = (np.random.randn(N, N) + 1j * np.random.randn(N, N)) / np.sqrt(2.0)
-    q, r = qr(z)
-    d = np.diag(r)
-    ph = d / np.abs(d)
-    U = np.multiply(q, ph, q)
-    return U
-
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_identity(r, tol):
-    """Tests the correct construction of the single mode identity operation"""
-    nmodes = 1
-    cutoff = 7
-    S = np.identity(2 * nmodes)
-    alphas = np.zeros([nmodes])
-    T = operations.fock_tensor(S, alphas, cutoff, r=r)
-    expected = np.identity(cutoff)
-    assert np.allclose(T, expected, atol=tol, rtol=0)
-
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_rotation(r, tol):
-    """Tests the correct construction of the single mode rotation operation"""
-    nmodes = 1
-    cutoff = 7
-    theta = 2 * np.pi * np.random.rand()
-    S = rotation(theta)
-    alphas = np.zeros([nmodes])
-    T = operations.fock_tensor(S, alphas, cutoff, r=r)
-    expected = np.diag(np.exp(1j * theta * np.arange(cutoff)))
-    assert np.allclose(T, expected, atol=tol, rtol=0)
-
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_displacement(r, tol):
-    """Tests the correct construction of the single mode displacement operation"""
-    nmodes = 1
-    cutoff = 5
-    alphas = (0.3 + 0.5 * 1j) * np.ones([nmodes])
-    S = np.identity(2 * nmodes)
[…]
