### 2019-11-12 631b1e9
removes operations.rst
