+++ docs/code/operations.rst
@@ -1,2 +1,0 @@
-.. automodule:: thewalrus.operations
-	:members:
