### 2019-11-12 6a48fe8
updates index.rst
