### 2019-11-12 9a94603
Fixes linting
