### 2019-11-12 33c62a2
Updates docs
