### 2019-11-13 95f0747
Apply suggestions from code review
Co-Authored-By: Josh Izaac <josh146@gmail.com>
