### 2019-11-13 259c854
Applies suggestions from code review
