### 2019-11-13 8f6a6a1
Moves sympmat and is_symplectic into the symplectic module
