### 2019-11-13 8853c42
Minor tweaks
