### 2019-11-13 10f2fd3
Get rid of Choi expand
