### 2019-11-13 905cf7c
Removes left over argument in two_mode_squeezing
