### 2019-11-13 2cc5b88
removes of unnecesary import of sympmat in test_quantum
