### 2019-11-13 02edf73
Updates notebook on fock states
