### 2019-11-17 301e95c
Improves coverage of reference
