### 2019-11-18 8e47bce
Added sf_ordering option in fock_tensor
