+++ thewalrus/tests/test_quantum.py
@@ -49,6 +49,7 @@ from thewalrus.quantum import (
 # make tests deterministic
 np.random.seed(137)
 
+
 def random_interferometer(N, real=False):
     r"""Random unitary matrix representing an interferometer.
 
@@ -81,12 +82,7 @@ def test_reduced_gaussian(n):
     cov = np.arange(N ** 2).reshape(N, N)
     res = reduced_gaussian(mu, cov, n)
     assert np.all(res[0] == np.array([n, n + m]))
-    assert np.all(
-        res[1]
-        == np.array(
-            [[(N + 1) * n, (N + 1) * n + m], [(N + 1) * n + N * m, (N + 1) * n + N * m + m]]
-        )
-    )
+    assert np.all(res[1] == np.array([[(N + 1) * n, (N + 1) * n + m], [(N + 1) * n + N * m, (N + 1) * n + N * m + m]]))
 
 
 def test_reduced_gaussian_two_mode():
@@ -125,7 +121,6 @@ def test_xmat(n):
     X = np.block([[O, I], [I, O]])
     res = Xmat(n)
     assert np.all(X == res)
-
 
 
 def test_Qmat_vacuum():
@@ -343,15 +338,7 @@ def test_density_matrix_squeezed():
 
     res = density_matrix(mu, V)
 
-    expected = np.array(
-        [
-            [0.91417429, 0, -0.26200733, 0, 0.09196943],
-            [0, 0, 0, 0, 0],
-            [-0.26200733, 0, 0.07509273, 0, -0.02635894],
-            [0, 0, 0, 0, 0],
-            [0.09196943, 0, -0.02635894, 0, 0.00925248],
-        ]
-    )
+    expected = np.array([[0.91417429, 0, -0.26200733, 0, 0.09196943], [0, 0, 0, 0, 0], [-0.26200733, 0, 0.07509273, 0, -0.02635894], [0, 0, 0, 0, 0], [0.09196943, 0, -0.02635894, 0, 0.00925248]])
 
     assert np.allclose(res, expected)
 
@@ -386,15 +373,7 @@ def test_density_matrix_squeezed_postselect():
 
     res = density_matrix(mu, V, post_select={0: 0}, cutoff=15, normalize=True)[:5, :5]
 
-    expected = np.array(
-        [
-            [0.91417429, 0, -0.26200733, 0, 0.09196943],
-            [0, 0, 0, 0, 0],
-            [-0.26200733, 0, 0.07509273, 0, -0.02635894],
-            [0, 0, 0, 0, 0],
-            [0.09196943, 0, -0.02635894, 0, 0.00925248],
-        ]
-    )
+    expected = np.array([[0.91417429, 0, -0.26200733, 0, 0.09196943], [0, 0, 0, 0, 0], [-0.26200733, 0, 0.07509273, 0, -0.02635894], [0, 0, 0, 0, 0], [0.09196943, 0, -0.02635894, 0, 0.00925248]])
 
     assert np.allclose(res, expected)
 
@@ -587,12 +566,7 @@ def test_state_vector_two_mode_squeezed():
     r = np.arcsinh(np.sqrt(nbar))
     cov = two_mode_squeezing(2 * r, phase)
     mu = np.zeros([4], dtype=np.complex)
-    exact = np.array(
-        [
-            (np.exp(1j * i * phase) * (nbar / (1.0 + nbar)) ** (i / 2) / np.sqrt(1.0 + nbar))
-            for i in range(cutoff)
-        ]
-    )
+    exact = np.array([(np.exp(1j * i * phase) * (nbar / (1.0 + nbar)) ** (i / 2) / np.sqrt(1.0 + nbar)) for i in range(cutoff)])
     psi = state_vector(mu, cov, cutoff=cutoff)
     expected = np.diag(exact)
[…]
