### 2019-11-18 0640eaf
Tests seems to pass
