+++ include/hermite_multidimensional.hpp
@@ -168,6 +168,11 @@ inline std::vector<T> hermite_multidimensional_cpp(std::vector<T> &R_mat, std::v
     std::vector<double> factors(resolution+1, 0);
     int jump = 0;
 
+    //    std::vector<T> local_y(dim, 0);
+    //    for (int kk = 0; kk < dim; kk++) {
+    //          local_y[kk] = 0.0;
+    //    }
+    
 
     for (ullint jj = 0; jj < Hdim-1; jj++) {
 
@@ -205,9 +210,10 @@ inline std::vector<T> hermite_multidimensional_cpp(std::vector<T> &R_mat, std::v
         ullint fromCoordinate = vec2index(jumpFrom, resolution);
 
 
-        for (int ii = 0; ii < dim; ii++) {
-            H[nextCoordinate] = H[nextCoordinate] + R(k, ii) * y(ii, 0);
-        }
+	//        for (int ii = 0; ii < dim; ii++) {
+	// H[nextCoordinate] = H[nextCoordinate] + R(k, ii) * y(ii, 0);
+	H[nextCoordinate] = H[nextCoordinate] + y(k, 0);
+	    //        }
         H[nextCoordinate] = H[nextCoordinate] * H[fromCoordinate];
 
         std::vector<int> tmpjump(dim, 0);
