+++ thewalrus/quantum.py
@@ -392,15 +392,14 @@ def density_matrix(mu, cov, post_select=None, normalize=False, cutoff=5, hbar=2)
         if np.allclose(mu, np.zeros_like(mu)):
             tensor = pref * hermite_multidimensional(-A, cutoff, renorm=True)
             return tensor.transpose(sf_order)
-        try:
-            beta = Beta(mu)
-            y = np.linalg.inv(A) @ (beta - A @ beta.conj())
-            tensor = pref * hermite_multidimensional(-A, cutoff, y=-y, renorm=True)
-            return tensor.transpose(sf_order)
+        
+        beta = Beta(mu)
+        y = (beta - A @ beta.conj())
+        tensor = pref * hermite_multidimensional(-A, cutoff, y=y, renorm=True)
+        return tensor.transpose(sf_order)
 
-        except np.linalg.LinAlgError:
-            pass
-        post_select = {}
+        
+        
 
     M = N - len(post_select)
     rho = np.zeros([cutoff] * (2 * M), dtype=np.complex128)
