### 2019-11-18 485a099
Removes comments in the C++ code
