### 2019-11-18 f6d9ee5
Skips the pool tests
