### 2019-11-18 fc08702
Skips the pool tests
