### 2019-11-18 8020511
Cleans up hafnian.ipynb and permanent.ipynb
