+++ docs/hafnian.bib
@@ -1,62 +1,0 @@
-@article{bjorklund2018faster,
-  title={A faster hafnian formula for complex matrices and its benchmarking on the titan supercomputer},
-  author={Bj{\"o}rklund, Andreas and Gupt, Brajesh and Quesada, Nicol{\'a}s},
-  journal={arXiv preprint arXiv:1805.12498},
-  year={2018}
-}
-
-@article{caianiello1953quantum,
-  title={On quantum field theory—I: explicit solution of Dyson’s equation in electrodynamics without use of feynman graphs},
-  author={Caianiello, Eduardo R},
-  journal={Il Nuovo Cimento (1943-1954)},
-  volume={10},
-  number={12},
-  pages={1634--1652},
-  year={1953},
-  publisher={Springer}
-}
-@article{banchi2019molecular,
-  title={Molecular Docking with Gaussian Boson Sampling},
-  author={Banchi, Leonardo and Fingerhuth, Mark and Babej, Tomas and Arrazola, Juan Miguel and others},
-  journal={arXiv preprint arXiv:1902.00462},
-  year={2019}
-}
-
-
-@inproceedings{bjorklund2012counting,
-  title={Counting perfect matchings as fast as Ryser},
-  author={Bj{\"o}rklund, Andreas},
-  booktitle={Proceedings of the twenty-third annual ACM-SIAM symposium on Discrete Algorithms},
-  pages={914--921},
-  year={2012},
-  organization={Society for Industrial and Applied Mathematics}
-}
-
-
-
-@book{barvinok2016combinatorics,
-  title={Combinatorics and complexity of partition functions},
-  author={Barvinok, Alexander},
-  volume={276},
-  year={2016},
-  publisher={Springer}
-}
-
-
-@article{quesada2019simulating,
-  title={Simulating realistic non-Gaussian state preparation},
-  author={Quesada, N and Helt, LG and Izaac, J and Arrazola, JM and Shahrokhshahi, R and Myers, CR and Sabapathy, KK},
-  journal={arXiv preprint arXiv:1905.07011},
-  year={2019}
-}
-
-@article{quesada2019franck,
-  title={Franck-Condon factors by counting perfect matchings of graphs with loops},
-  author={Quesada, Nicol{\'a}s},
-  journal={The Journal of chemical physics},
-  volume={150},
-  number={16},
-  pages={164113},
-  year={2019},
-  publisher={AIP Publishing}
-}
