### 2019-11-18 b0ba1f9
Correct errors in tests and improves coverage
