+++ thewalrus/tests/test_quantum.py
@@ -634,12 +634,11 @@ def test_is_classical_cov_thermal(nbar):
 
 
 @pytest.mark.parametrize("cutoff", [50, 51, 52, 53])
-def test_total_photon_num_dist_pure_state(cutoff):
+def test_total_photon_num_dist_pure_state(cutoff, atol=1e-07):
     """ Test the correct photon number distribution is obtained for n modes
     with nmean number of photons up to Fock cutoff nmax"""
     n = 3
     nmean = 1.0
-    cutoff = 50
     rs = np.arcsinh(np.sqrt(nmean)) * np.ones([n])
     cov = np.diag(np.concatenate([np.exp(2 * rs), np.exp(-2 * rs)]))
     p1 = total_photon_num_dist_pure_state(cov, cutoff=cutoff)
