### 2019-11-18 6fbb1f6
removes unused argument
