+++ thewalrus/tests/test_quantum.py
@@ -634,7 +634,7 @@ def test_is_classical_cov_thermal(nbar):
 
 
 @pytest.mark.parametrize("cutoff", [50, 51, 52, 53])
-def test_total_photon_num_dist_pure_state(cutoff, atol=1e-07):
+def test_total_photon_num_dist_pure_state(cutoff):
     """ Test the correct photon number distribution is obtained for n modes
     with nmean number of photons up to Fock cutoff nmax"""
     n = 3
