### 2019-11-18 a2fb4bd
recreates libwalrus.cpp
