### 2019-11-18 c47ba06
fixes linting issues
