### 2019-11-18 4ac095c
fixes linting issues
