### 2019-11-18 e0fcfd1
fixes linting issues
