### 2019-11-18 841589d
Removes commented lines and print statements
