### 2019-11-19 cee0a52
Before deleting unused functions
