+++ thewalrus/_hermite_multidimensional.py
@@ -21,7 +21,7 @@ from .libwalrus import hermite_multidimensional as hm
 from ._hafnian import input_validation
 
 
-def hermite_multidimensional(R, cutoff, y=None, renorm=False, make_tensor=True):
+def hermite_multidimensional(R, cutoff, y=None, renorm=False, make_tensor=True, modified=False):
     r"""Returns the multidimensional Hermite polynomials :math:`H_k^{(R)}(y)`.
 
     Here :math:`R` is an :math:`n \times n` square matrix, and
@@ -46,14 +46,25 @@ def hermite_multidimensional(R, cutoff, y=None, renorm=False, make_tensor=True):
         y (array): vector argument of the Hermite polynomial
         renorm (bool): If ``True``, normalizes the returned multidimensional Hermite
             polynomials such that :math:`H_k^{(R)}(y)/\prod_i k_i!`
-        make_tensor: If ``False``, returns a flattened one dimensional array
+        make_tensor (bool): If ``False``, returns a flattened one dimensional array
             containing the values of the polynomial
+        modified (bool): whether to return the modified multidimensional Hermite polynomials or the standard ones
 
     Returns:
         (array): the multidimensional Hermite polynomials
     """
+
     input_validation(R)
     n, _ = R.shape
+
+    if modified == False and y is not None:
+        m = y.shape[0]
+        if m == n:
+            ym = R @ y
+            return hermite_multidimensional(
+                R, cutoff, y=ym, renorm=renorm, make_tensor=make_tensor, modified=True
+            )
+
     if y is None:
         y = np.zeros([n], dtype=complex)
 
@@ -107,11 +118,15 @@ def hafnian_batched(A, cutoff, mu=None, tol=1e-12, renorm=False, make_tensor=Tru
     input_validation(A, tol=tol)
     n, _ = A.shape
 
-    #if not np.allclose(A, np.zeros([n, n])):
+    # if not np.allclose(A, np.zeros([n, n])):
     if mu is not None:
         return hermite_multidimensional(
-            -A, cutoff, y=mu, renorm=renorm, make_tensor=make_tensor
+            -A, cutoff, y=mu, renorm=renorm, make_tensor=make_tensor, modified=True
         )
     yi = np.zeros([n], dtype=complex)
-    return hermite_multidimensional(-A, cutoff, y=yi, renorm=renorm, make_tensor=make_tensor)
+    return hermite_multidimensional(
+        -A, cutoff, y=yi, renorm=renorm, make_tensor=make_tensor, modified=True
+    )
+
+
 # Note the minus signs in the arguments. Those are intentional and are due to the fact that Dodonov et al. in PRA 50, 813 (1994) use (p,q) ordering instead of (q,p) ordering
