+++ thewalrus/quantum.py
@@ -380,11 +380,11 @@ def density_matrix(mu, cov, post_select=None, normalize=False, cutoff=5, hbar=2)
         sf_order = tuple(chain.from_iterable([[i, i + N] for i in range(N)]))
 
         if np.allclose(mu, np.zeros_like(mu)):
-            tensor = pref * hermite_multidimensional(-A, cutoff, renorm=True)
+            tensor = pref * hermite_multidimensional(-A, cutoff, renorm=True, modified=True)
             return tensor.transpose(sf_order)
         beta = Beta(mu)
         y = beta - A @ beta.conj()
-        tensor = pref * hermite_multidimensional(-A, cutoff, y=y, renorm=True)
+        tensor = pref * hermite_multidimensional(-A, cutoff, y=y, renorm=True, modified=True)
         return tensor.transpose(sf_order)
 
     M = N - len(post_select)
