### 2019-11-19 addeba1
Fixes linting issues
