### 2019-11-19 2592890
Fixes linting issues
