### 2019-11-19 a40bdc1
Applies code review suggestions
