### 2019-11-19 70241ac
Removes commented line
