### 2019-11-20 7acdf73
Updates quick_guide [ci skip]
