### 2019-11-20 89408f2
Updates quick_guide [ci skip]
