### 2019-11-20 4fdf902
Adds test to check relation between standard hermite and modified hermite
