### 2019-12-06 32960f1
Adds X and Z gates
