### 2019-12-06 1d85c43
All the gates compile and differentiate correctly
