### 2019-12-06 b2e9d56
Adds docstring
