+++ thewalrus/fock_gradients.py
@@ -60,13 +60,24 @@ import numpy as np
 
 # @numba.jit
 def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
+    r"""
+    Calculates the Fock representation of the Xgate and its gradient
+    Arg:
+        x (float): Parameter of the gate
+        cutoff (int): Fock ladder cutoff
+        grad (boolean): Whether to calculate the gradient or not
+        hbar (float): Value of hbar is the commutation relation
+        r (float): Value of the parameter used internally in fock_tensor. 
+    Returns:
+        tuple(G,dG): The Fock representations of the gate and its gradient
+    """
     nmodes = 1
     pref = 1.0 / np.sqrt(2 * hbar)
     alpha = np.array([pref * x])
     S = np.identity(2 * nmodes)
 
     if not grad:
-        return fock_tensor(S, alpha, cutoff, r=r)
+        return fock_tensor(S, alpha, cutoff, r=r), None
 
     T = fock_tensor(S, alpha, cutoff + 1, r=r)
     gradT = np.zeros([cutoff, cutoff], dtype=complex)
@@ -80,13 +91,24 @@ def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
 
 # @numba.jit
 def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
+    r"""
+    Calculates the Fock representation of the Zgate and its gradient
+    Arg:
+        p (float): Parameter of the gate
+        cutoff (int): Fock ladder cutoff
+        grad (boolean): Whether to calculate the gradient or not
+        hbar (float): Value of hbar is the commutation relation
+        r (float): Value of the parameter used internally in fock_tensor.
+    Returns:
+        tuple(G,dG): The Fock representations of the gate and its gradient
+    """
     nmodes = 1
     pref = 1.0 / np.sqrt(2 * hbar)
     alpha = np.array([1j * pref * p])
     S = np.identity(2 * nmodes)
 
     if not grad:
-        return fock_tensor(S, alpha, cutoff, r=r)
+        return fock_tensor(S, alpha, cutoff, r=r), None
 
     T = fock_tensor(S, alpha, cutoff + 1, r=r)
     gradT = np.zeros([cutoff, cutoff], dtype=complex)
@@ -100,9 +122,19 @@ def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
 
 # @numba.jit
 def Sgate(s, cutoff, grad=False, r=np.arcsinh(1.0)):
+    r"""
+    Calculates the Fock representation of the Sgate and its gradient
+    Arg:
+        s (float): Parameter of the gate
+        cutoff (int): Fock ladder cutoff
+        grad (boolean): Whether to calculate the gradient or not
+        r (float): Value of the parameter used internally in fock_tensor.
+    Returns:
+        tuple(G,dG): The Fock representations of the gate and its gradient
+    """
     S = squeezing(s, 0.0)
     if not grad:
-        return fock_tensor(S, np.zeros([1]), cutoff, r=r).real
+        return fock_tensor(S, np.zeros([1]), cutoff, r=r).real, None
 
     T = fock_tensor(S, np.zeros([1]), cutoff + 2, r=r).real
     gradT = np.zeros([cutoff, cutoff])
@@ -116,17 +148,36 @@ def Sgate(s, cutoff, grad=False, r=np.arcsinh(1.0)):
 
 
 def Rgate(theta, cutoff, grad=False):
+    r"""
+    Calculates the Fock representation of the Rgate and its gradient
+    Arg:
+        s (float): Parameter of the gate
+        cutoff (int): Fock ladder cutoff
[…]
