### 2019-12-06 5b49e20
Adds tests
