### 2019-12-06 38e1277
Fixes linting issues
