### 2019-12-06 d1efea6
Adds docstrings
