### 2019-12-07 94ad190
Corrects typos
