+++ thewalrus/libwalrus.cpp
@@ -1,4 +1,4 @@
-/* Generated by Cython 0.29.2 */
+/* Generated by Cython 0.29.6 */
 
 /* BEGIN: Cython Metadata
 {
@@ -22,7 +22,7 @@
             "-fPIC",
             "-shared",
             "-fopenmp",
-            "-I/home/josh/.local/lib/python3.6/site-packages/numpy/core/include"
+            "-I/home/nquesada/miniconda3/envs/sf/lib/python3.6/site-packages/numpy/core/include"
         ],
         "extra_link_args": [
             "-fopenmp"
@@ -30,7 +30,7 @@
         "include_dirs": [
             "thewalrus",
             "",
-            "/home/josh/.local/lib/python3.6/site-packages/numpy/core/include",
+            "/home/nquesada/miniconda3/envs/sf/lib/python3.6/site-packages/numpy/core/include",
             "/usr/local/include/eigen3",
             "/usr/include/eigen3",
             "include"
@@ -38,8 +38,7 @@
         "language": "c++",
         "library_dirs": [
             "/usr/lib",
-            "/usr/local/lib",
-            "/home/josh/.local/lib"
+            "/usr/local/lib"
         ],
         "name": "libwalrus",
         "sources": [
@@ -57,8 +56,8 @@ END: Cython Metadata */
 #elif PY_VERSION_HEX < 0x02060000 || (0x03000000 <= PY_VERSION_HEX && PY_VERSION_HEX < 0x03030000)
     #error Cython requires Python 2.6+ or Python 3.3+.
 #else
-#define CYTHON_ABI "0_29_2"
-#define CYTHON_HEX_VERSION 0x001D02F0
+#define CYTHON_ABI "0_29_6"
+#define CYTHON_HEX_VERSION 0x001D06F0
 #define CYTHON_FUTURE_DIVISION 0
 #include <stddef.h>
 #ifndef offsetof
@@ -410,26 +409,6 @@ class __Pyx_FakeReference {
 #else
 #define __Pyx_PyFastCFunction_Check(func) 0
 #endif
-#if CYTHON_USE_DICT_VERSIONS
-#define __PYX_GET_DICT_VERSION(dict)  (((PyDictObject*)(dict))->ma_version_tag)
-#define __PYX_UPDATE_DICT_CACHE(dict, value, cache_var, version_var)\
-    (version_var) = __PYX_GET_DICT_VERSION(dict);\
-    (cache_var) = (value);
-#define __PYX_PY_DICT_LOOKUP_IF_MODIFIED(VAR, DICT, LOOKUP) {\
-        static PY_UINT64_T __pyx_dict_version = 0;\
-        static PyObject *__pyx_dict_cached_value = NULL;\
-        if (likely(__PYX_GET_DICT_VERSION(DICT) == __pyx_dict_version)) {\
-            (VAR) = __pyx_dict_cached_value;\
-        } else {\
-            (VAR) = __pyx_dict_cached_value = (LOOKUP);\
-            __pyx_dict_version = __PYX_GET_DICT_VERSION(DICT);\
-        }\
-    }
-#else
-#define __PYX_GET_DICT_VERSION(dict)  (0)
-#define __PYX_UPDATE_DICT_CACHE(dict, value, cache_var, version_var)
-#define __PYX_PY_DICT_LOOKUP_IF_MODIFIED(VAR, DICT, LOOKUP)  (VAR) = (LOOKUP);
-#endif
 #if CYTHON_COMPILING_IN_PYPY && !defined(PyObject_Malloc)
[…]
