### 2019-12-08 a65b1dd
Solves linting
