+++ thewalrus/fock_gradients.py
@@ -37,10 +37,11 @@ Code details
 ^^^^^^^^^^^^
 """
 import numpy as np
-from numba import double, jit 
+from numba import double, jit
 
 from thewalrus.quantum import fock_tensor
 from thewalrus.symplectic import squeezing, two_mode_squeezing, beam_splitter
+
 
 def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
     r"""
@@ -72,7 +73,7 @@ def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
     return T[0:cutoff, 0:cutoff], gradT
 
 
-#@numba.jit
+# @numba.jit
 def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
     r"""
     Calculates the Fock representation of the Zgate and its gradient
@@ -103,7 +104,7 @@ def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
     return T[0:cutoff, 0:cutoff], gradT
 
 
-#@numba.jit
+# @numba.jit
 def Sgate(s, cutoff, grad=False, r=np.arcsinh(1.0)):
     r"""
     Calculates the Fock representation of the Sgate and its gradient
@@ -146,7 +147,8 @@ def Rgate(theta, cutoff, grad=False):
         return np.diag(T), None
     return np.diag(T), np.diag(1j * ns * T)
 
-#@numba.jit(nopython=True)
+
+# @numba.jit(nopython=True)
 def grad_S2gate(T, cutoff):
     """Ready for numba"""
     gradT = np.zeros([cutoff, cutoff, cutoff, cutoff])
@@ -159,6 +161,7 @@ def grad_S2gate(T, cutoff):
                     if m > 0 and l > 0:
                         gradT[n, k, m, l] -= np.sqrt(m * l) * T[n, k, m - 1, l - 1]
     return gradT
+
 
 def S2gate(s, cutoff, grad=False, r=np.arcsinh(1.0)):
     r"""
@@ -176,10 +179,11 @@ def S2gate(s, cutoff, grad=False, r=np.arcsinh(1.0)):
         return fock_tensor(S, np.zeros([2]), cutoff, r=r).real, None
 
     T = fock_tensor(S, np.zeros([2]), cutoff + 1, r=r).real
-    grad_S2gatejit = jit(double[:,:,:,:](double[:,:,:,:], double))(grad_S2gate) 
+    grad_S2gatejit = jit(double[:, :, :, :](double[:, :, :, :], double))(grad_S2gate)
     return T[0:cutoff, 0:cutoff, 0:cutoff, 0:cutoff], grad_S2gatejit(T, cutoff)
 
-#@numba.jit
+
+# @numba.jit
 def BSgate(theta, cutoff, grad=False, r=np.arcsinh(1.0)):
     r"""
     Calculates the Fock representation of the BSgate and its gradient
