### 2019-12-09 da22429
All functions work with numba
