+++ thewalrus/fock_gradients.py
@@ -37,10 +37,28 @@ Code details
 ^^^^^^^^^^^^
 """
 import numpy as np
-from numba import double, jit
+from numba import jit, double, complex128
 
 from thewalrus.quantum import fock_tensor
 from thewalrus.symplectic import squeezing, two_mode_squeezing, beam_splitter
+
+@jit('double[:,:](double[:,:], double)')
+def grad_Xgate(T, pref):
+    """Calculate the gradient of the Xgate
+    Args:
+        T (array): Array representing the gate
+    Returns:
+        (array): Array representing the gradient of the gate
+    """
+    cutoff, _ = T.shape
+    cutoff -= 1
+    gradT = np.zeros([cutoff, cutoff])
+    for n in range(cutoff):
+        for m in range(cutoff):
+            gradT[n, m] = np.sqrt(m + 1) * T[n, m + 1] * pref
+            if m > 0:
+                gradT[n, m] -= np.sqrt(m) * T[n, m - 1] * pref
+    return gradT
 
 
 def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
@@ -64,16 +82,27 @@ def Xgate(x, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
         return fock_tensor(S, alpha, cutoff, r=r).real, None
 
     T = fock_tensor(S, alpha, cutoff + 1, r=r).real
-    gradT = np.zeros([cutoff, cutoff])
+    return T[0:cutoff, 0:cutoff], grad_Xgate(T, pref)
+
+
+@jit('complex128[:,:](complex128[:,:], double)')
+def grad_Zgate(T, pref):
+    """Calculate the gradient of the Zgate
+    Args:
+        T (array): Array representing the gate
+    Returns:
+        (array): Array representing the gradient of the gate
+    """
+    cutoff, _ = T.shape
+    cutoff -= 1
+    gradT = np.zeros([cutoff, cutoff], dtype=complex)
     for n in range(cutoff):
         for m in range(cutoff):
-            gradT[n, m] = np.sqrt(m + 1) * T[n, m + 1] * pref
+            gradT[n, m] = 1j * np.sqrt(m + 1) * T[n, m + 1] * pref
             if m > 0:
-                gradT[n, m] -= np.sqrt(m) * T[n, m - 1] * pref
-    return T[0:cutoff, 0:cutoff], gradT
+                gradT[n, m] += 1j * np.sqrt(m) * T[n, m - 1] * pref
+    return gradT
 
-
-# @numba.jit
 def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
     r"""
     Calculates the Fock representation of the Zgate and its gradient
@@ -95,16 +124,29 @@ def Zgate(p, cutoff, grad=False, hbar=2, r=np.arcsinh(1.0)):
         return fock_tensor(S, alpha, cutoff, r=r), None
 
     T = fock_tensor(S, alpha, cutoff + 1, r=r)
-    gradT = np.zeros([cutoff, cutoff], dtype=complex)
+    return T[0:cutoff, 0:cutoff], grad_Zgate(T, pref)
+
+@jit('double[:,:](double[:,:])')
+def grad_Sgate(T):
+    """Calculate the gradient of the Sgate
+    Args:
+        T (array): Array representing the gate
+    Returns:
+        (array): Array representing the gradient of the gate
+    """
+    cutoff, _ = T.shape
+    cutoff -= 2
+    gradT = np.zeros([cutoff, cutoff])
     for n in range(cutoff):
-        for m in range(cutoff):
[…]
