### 2019-12-09 be22e3b
Now hermite tries to do real calculation if inputs are real
