+++ thewalrus/_hermite_multidimensional.py
@@ -17,6 +17,7 @@ Hermite Multidimensional Python interface
 import numpy as np
 
 from .libwalrus import hermite_multidimensional as hm
+from .libwalrus import hermite_multidimensional as hmr
 from ._hafnian import input_validation
 
 # pylint: disable=too-many-arguments
@@ -71,13 +72,20 @@ def hermite_multidimensional(R, cutoff, y=None, renorm=False, make_tensor=True, 
     if m != n:
         raise ValueError("The matrix R and vector y have incompatible dimensions")
 
-    values = np.array(hm(R, y, cutoff, renorm=renorm))
+    Rt = np.real_if_close(R)
+    yt = np.real_if_close(y)
+
+    if Rt.dtype == np.float and yt.dtype == np.float:
+        values = np.array(hmr(R, y, cutoff, renorm=renorm))
+    else:
+        values = np.array(hm(R, y, cutoff, renorm=renorm))
 
     if make_tensor:
         shape = cutoff * np.ones([n], dtype=int)
         values = np.reshape(values, shape)
 
     return values
+
 
 # pylint: disable=too-many-arguments
 def hafnian_batched(A, cutoff, mu=None, tol=1e-12, renorm=False, make_tensor=True):
@@ -104,7 +112,7 @@ def hafnian_batched(A, cutoff, mu=None, tol=1e-12, renorm=False, make_tensor=Tru
         cutoff (int): maximum size of the subindices in the Hermite polynomial
         mu (array): a vector of length :math:`N` representing the vector of means/displacement
         renorm (bool): If ``True``, the returned hafnians are *normalized*, that is,
-            :math:`haf(reduction(A, k))/\prod_i k_i!`
+            :math:`haf(reduction(A, k))/\ \sqrt{prod_i k_i!}`
             (or :math:`lhaf(fill\_diagonal(reduction(A, k),reduction(mu, k)))` if
             ``mu`` is not None)
         make_tensor: If ``False``, returns a flattened one dimensional array instead
@@ -116,13 +124,11 @@ def hafnian_batched(A, cutoff, mu=None, tol=1e-12, renorm=False, make_tensor=Tru
     input_validation(A, tol=tol)
     n, _ = A.shape
 
-    if mu is not None:
-        return hermite_multidimensional(
-            -A, cutoff, y=mu, renorm=renorm, make_tensor=make_tensor, modified=True
-        )
-    yi = np.zeros([n], dtype=complex)
+    if mu is None:
+        mu = np.zeros([n], dtype=complex)
+
     return hermite_multidimensional(
-        -A, cutoff, y=yi, renorm=renorm, make_tensor=make_tensor, modified=True
+        -A, cutoff, y=mu, renorm=renorm, make_tensor=make_tensor, modified=True
     )
 
 
