### 2019-12-09 241b353
Recreated libwalrus.cpp
