### 2019-12-09 c344f07
Updates docstring and gradients work using numba
