+++ thewalrus/libwalrus.cpp
@@ -1,4 +1,4 @@
-/* Generated by Cython 0.29.2 */
+/* Generated by Cython 0.29.4 */
 
 /* BEGIN: Cython Metadata
 {
@@ -22,7 +22,7 @@
             "-fPIC",
             "-shared",
             "-fopenmp",
-            "-I/home/josh/.local/lib/python3.6/site-packages/numpy/core/include"
+            "-I/home/nicolas/.local/lib/python3.6/site-packages/numpy/core/include"
         ],
         "extra_link_args": [
             "-fopenmp"
@@ -30,7 +30,7 @@
         "include_dirs": [
             "thewalrus",
             "",
-            "/home/josh/.local/lib/python3.6/site-packages/numpy/core/include",
+            "/home/nicolas/.local/lib/python3.6/site-packages/numpy/core/include",
             "/usr/local/include/eigen3",
             "/usr/include/eigen3",
             "include"
@@ -38,8 +38,7 @@
         "language": "c++",
         "library_dirs": [
             "/usr/lib",
-            "/usr/local/lib",
-            "/home/josh/.local/lib"
+            "/usr/local/lib"
         ],
         "name": "libwalrus",
         "sources": [
@@ -57,8 +56,8 @@ END: Cython Metadata */
 #elif PY_VERSION_HEX < 0x02060000 || (0x03000000 <= PY_VERSION_HEX && PY_VERSION_HEX < 0x03030000)
     #error Cython requires Python 2.6+ or Python 3.3+.
 #else
-#define CYTHON_ABI "0_29_2"
-#define CYTHON_HEX_VERSION 0x001D02F0
+#define CYTHON_ABI "0_29_4"
+#define CYTHON_HEX_VERSION 0x001D04F0
 #define CYTHON_FUTURE_DIVISION 0
 #include <stddef.h>
 #ifndef offsetof
@@ -462,7 +461,7 @@ class __Pyx_FakeReference {
 typedef int Py_tss_t;
 static CYTHON_INLINE int PyThread_tss_create(Py_tss_t *key) {
   *key = PyThread_create_key();
-  return 0; // PyThread_create_key reports success always
+  return 0;
 }
 static CYTHON_INLINE Py_tss_t * PyThread_tss_alloc(void) {
   Py_tss_t *key = (Py_tss_t *)PyObject_Malloc(sizeof(Py_tss_t));
@@ -485,7 +484,7 @@ static CYTHON_INLINE int PyThread_tss_set(Py_tss_t *key, void *value) {
 static CYTHON_INLINE void * PyThread_tss_get(Py_tss_t *key) {
   return PyThread_get_key_value(*key);
 }
-#endif // TSS (Thread Specific Storage) API
+#endif
 #if CYTHON_COMPILING_IN_CPYTHON || defined(_PyDict_NewPresized)
 #define __Pyx_PyDict_NewPresized(n)  ((n <= 8) ? PyDict_New() : _PyDict_NewPresized(n))
 #else
@@ -1564,9 +1563,9 @@ static CYTHON_INLINE int __Pyx_PyErr_GivenExceptionMatches2(PyObject *err, PyObj
 static CYTHON_UNUSED int __pyx_memoryview_getbuffer(PyObject *__pyx_v_self, Py_buffer *__pyx_v_info, int __pyx_v_flags); /*proto*/
 /* PyIntBinop.proto */
 #if !CYTHON_COMPILING_IN_PYPY
-static PyObject* __Pyx_PyInt_AddObjC(PyObject *op1, PyObject *op2, long intval, int inplace);
[…]
