### 2019-12-09 787dd00
Removes commented line
