### 2019-12-09 7015cde
Adds pragma no cover to the jit functions. They are still being tested by calling the gates functions themselves
