### 2019-12-10 2daa14b
Adds two parameter gates
