### 2019-12-10 d044a9d
The gradient for BS still broken
