+++ thewalrus/fock_gradients2.py
@@ -44,8 +44,8 @@ from thewalrus.quantum import fock_tensor
 from thewalrus.symplectic import squeezing, two_mode_squeezing, beam_splitter
 
 
-@jit("void(complex128[:,:], complex128[:,:], complex128[:,:], complex128, complex128)")
-def grad_Dgate(T, gradTr, gradTtheta, rexptheta, exptheta):# pragma: no cover
+@jit("void(complex128[:,:], complex128[:,:], complex128[:,:], double)")
+def grad_Dgate(T, gradTr, gradTtheta, theta):  # pragma: no cover
     """Calculates the gradient of the Xgate.
 
     Args:
@@ -55,23 +55,21 @@ def grad_Dgate(T, gradTr, gradTtheta, rexptheta, exptheta):# pragma: no cover
     """
 
     cutoff = gradTr.shape[0]
+    exptheta = np.exp(1j * theta)
     for n in range(cutoff):
         for m in range(cutoff):
-            sqm1 = np.sqrt(m + 1)
-            sqm = np.sqrt(m)
-            gradTtheta[n, m] = 1j*(np.abs(rexptheta)**2)*T[n, m]
-            gradTr[n, m] =  sqm1 * T[n, m + 1] * exptheta
-            gradTtheta[n, m] += 1j * sqm1 * T[n, m + 1] * rexptheta
+            gradTtheta[n, m] = 1j * (n - m) * T[n, m]
+            gradTr[n, m] = np.sqrt(m + 1) * T[n, m + 1] * exptheta
             if m > 0:
-                gradTr[n, m] -= sqm * T[n, m - 1] * np.conj(exptheta)
-                gradTtheta[n, m] += 1j * sqm * T[n, m - 1] * np.conj(rexptheta)
+                gradTr[n, m] -= np.sqrt(m) * T[n, m - 1] * np.conj(exptheta)
 
 
 def Dgate(r, theta, cutoff, grad=False, hbar=2, s=np.arcsinh(1.0)):
     """Calculates the Fock representation of the Xgate and its gradient.
 
     Arg:
-        x (float): parameter of the gate
+        r (float): magnitude parameter of the gate
+        theta (float): magnitude parameter of the gate
         cutoff (int): Fock ladder cutoff
         grad (boolean): whether to calculate the gradient or not
         hbar (float): value of hbar in the commutation relation
@@ -81,7 +79,7 @@ def Dgate(r, theta, cutoff, grad=False, hbar=2, s=np.arcsinh(1.0)):
         tuple[array[float], array[float] or None]: The Fock representations of the gate and its gradient with size ``[cutoff]*2``
     """
     nmodes = 1
-    alpha = np.array([r * np.exp(1j*theta)])
+    alpha = np.array([r * np.exp(1j * theta)])
     S = np.identity(2 * nmodes)
 
     if not grad:
@@ -90,8 +88,149 @@ def Dgate(r, theta, cutoff, grad=False, hbar=2, s=np.arcsinh(1.0)):
     T = fock_tensor(S, alpha, cutoff + 1, r=s)
     gradTr = np.zeros([cutoff, cutoff], dtype=complex)
     gradTtheta = np.zeros([cutoff, cutoff], dtype=complex)
-    rexptheta = r*np.exp(1j*theta)
-    exptheta = np.exp(1j*theta)
-    grad_Dgate(T, gradTr, gradTtheta, rexptheta, exptheta)
+    grad_Dgate(T, gradTr, gradTtheta, theta)
     return T[0:cutoff, 0:cutoff], gradTr, gradTtheta
+
+
+
+
+
+@jit("void(complex128[:,:], complex128[:,:], complex128[:,:], double)")
[…]
