### 2019-12-11 d6349fc
Moves all the gates to fock_gradient
