+++ thewalrus/tests/test_fock_gradients.py
@@ -13,48 +13,120 @@
 # limitations under the License.
 """Tests for The Walrus fock_gradients functions"""
 
-from thewalrus.fock_gradients import Xgate, Zgate, Sgate, S2gate, BSgate, Rgate
+from thewalrus.fock_gradients import Dgate, BSgate, Sgate, S2gate, Xgate_one_param, Zgate_one_param, Sgate_one_param, S2gate_one_param, BSgate_one_param, Rgate
 import numpy as np
 
-
-def test_Xgate():
-    """Tests the value of the analytic gradient for the Xgate against finite differences"""
-    cutoff = 10
-    x = 1.0
-    _, dX = Xgate(x, cutoff, grad=True)
-    dx = 0.001
-    Xp, _ = Xgate(x + dx, cutoff)
-    Xm, _ = Xgate(x - dx, cutoff)
-    dXfd = (Xp - Xm) / (2 * dx)
-    assert np.allclose(dX, dXfd, atol=1e-5, rtol=0)
-
-
-def test_Zgate():
-    """Tests the value of the analytic gradient for the Zgate against finite differences"""
-    cutoff = 10
-    p = 1.0
-    _, dZ = Zgate(p, cutoff, grad=True)
-    dp = 0.001
-    Zp, _ = Zgate(p + dp, cutoff)
-    Zm, _ = Zgate(p - dp, cutoff)
-    dZfd = (Zp - Zm) / (2 * dp)
-    assert np.allclose(dZ, dZfd, atol=1e-5, rtol=0)
+def test_Dgate():
+    """Tests the value of the analytic gradient for the Dgate against finite differences"""
+    cutoff = 4
+    r = 1.0
+    theta = np.pi / 8
+    _, Dr, Dtheta = Dgate(r, theta, cutoff, grad=True)
+    dr = 0.001
+    dtheta = 0.001
+    Drp, _, _ = Dgate(r + dr, theta, cutoff, grad=False)
+    Drm, _, _ = Dgate(r - dr, theta, cutoff, grad=False)
+    Dthetap, _, _ = Dgate(r, theta + dtheta, cutoff, grad=False)
+    Dthetam, _, _ = Dgate(r, theta - dtheta, cutoff, grad=False)
+    Drapprox = (Drp - Drm) / (2 * dr)
+    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
+    assert np.allclose(Dr, Drapprox, atol=1e-5, rtol=0)
+    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-5, rtol=0)
 
 
 def test_Sgate():
     """Tests the value of the analytic gradient for the Sgate against finite differences"""
+    cutoff = 4
+    r = 1.0
+    theta = np.pi / 8
+    _, Dr, Dtheta = Sgate(r, theta, cutoff, grad=True)
+    dr = 0.001
+    dtheta = 0.001
+    Drp, _, _ = Sgate(r + dr, theta, cutoff, grad=False)
+    Drm, _, _ = Sgate(r - dr, theta, cutoff, grad=False)
+    Dthetap, _, _ = Sgate(r, theta + dtheta, cutoff, grad=False)
+    Dthetam, _, _ = Sgate(r, theta - dtheta, cutoff, grad=False)
+    Drapprox = (Drp - Drm) / (2 * dr)
+    Dthetaapprox = (Dthetap - Dthetam) / (2 * dtheta)
+    assert np.allclose(Dr, Drapprox, atol=1e-5, rtol=0)
+    assert np.allclose(Dtheta, Dthetaapprox, atol=1e-5, rtol=0)
+
+
+def test_S2gate():
[…]
