### 2019-12-11 0c83ac6
Moves all functions into a single file
