+++ thewalrus/fock_gradients2.py
@@ -1,43 +1,0 @@
-# Copyright 2019 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Gradients of Gaussian gates in the Fock representation
-======================================================
-
-**Module name:** :mod:`thewalrus.fock_gradients`
-
-.. currentmodule:: thewalrus.fock_gradients
-
-Contains the Fock representation of the standard Gaussian gates, as well as their gradients.
-
-
-Fock Gates
-----------
-
-.. autosummary::
-    Dgate
-    Sgate
-    S2gate
-    BSgate
-
-Code details
-^^^^^^^^^^^^
-"""
-import numpy as np
-
-from numba import jit
-
-from thewalrus.quantum import fock_tensor
-from thewalrus.symplectic import squeezing, two_mode_squeezing, beam_splitter
-
