### 2019-12-11 80fa21b
Bring back the old libwalrus.cpp
