### 2019-12-11 e5fdbaf
Correct bug by making sure the arguments of numba functions for two param gates are always complex
