### 2019-12-11 c453050
np.complex -> np.complex128
