### 2019-12-11 16a2fff
Corrects minor typos
