### 2019-12-11 d7ec586
Fixes docstrings
