+++ thewalrus/fock_gradients.py
@@ -74,7 +74,7 @@ def Dgate(r, theta, cutoff, grad=False, s=np.arcsinh(1.0)):
 
     Arg:
         r (float): displacement magnitude
-        theta (float): magnitude parameter of the gate
+        theta (float): displacement phase
         cutoff (int): Fock ladder cutoff
         grad (boolean): whether to calculate the gradient or not
         s (float): value of the parameter used internally in fock_tensor
