+++ thewalrus/fock_gradients.py
@@ -121,7 +121,7 @@ def Sgate(r, theta, cutoff, grad=False, s=np.arcsinh(1.0)):
     """Calculates the Fock representation of the Sgate and its gradient.
 
     Arg:
-        r (float): magnitude parameter of the gate
+        r (float): squeezing magnitude
         theta (float): magnitude parameter of the gate
         cutoff (int): Fock ladder cutoff
         grad (boolean): whether to calculate the gradient or not
