+++ docs/code/fock_gradients.rst
@@ -1,0 +1,3 @@
+.. automodule:: thewalrus.fock_gradients
+	:members:
+	:exclude-members: Dgate, Kgate
