+++ docs/code.rst
@@ -12,6 +12,10 @@ Python interface
 
 * The :mod:`thewalrus.samples` submodule provides access to algorithms to sample from the hafnian or the torontonian of Gaussian quantum states
 
+* The :mod:`thewalrus.symplectic` submodule provides access to a convenient set of symplectic matrices and utility functions to manipulate them
+
+* The :mod:`thewalrus.fock_gradients` submodule provides access to the Fock representation of certain continuous-variable gates and their gradients
+
 * The :mod:`thewalrus.reference` submodule provides access to pure-Python reference implementations of the hafnian, loop hafnian, and torontonian
 
 
