### 2019-12-12 c6a3367
Incorporates Josh's code review
