+++ thewalrus/fock_gradients.py
@@ -21,6 +21,10 @@ Gradients of Gaussian gates in the Fock representation
 
 Contains the Fock representation of the standard Gaussian gates and the Kerr gate, as well as their gradients.
 
+All the functions have an optional parameter called `choi_r` which can be used to set the value of the two-mode squeezed
+states used in the Choi-Jamiolkowski expansion of a given Gaussian gate. For details on this see the documentation of
+:mod:`thewalrus.quantum.fock_tensor`.
+
 
 Fock Gates
 ----------
@@ -55,9 +59,9 @@ def grad_Dgate(T, gradTr, gradTtheta, theta):  # pragma: no cover
 
     Args:
         T (array[complex]): array representing the gate
-        gradTr (array[complex]): array of zeros that will contain the value of the gradient with respect to r
-        gradTtheta (array[complex]): array of zeros that will contain the value of the gradient with respect to theta
-        theta (float): phase angle parametrizing the gate
+        gradTr (array[complex]): array of zeros that will contain the value of the gradient with respect to r the displacement magnitude
+        gradTtheta (array[complex]): array of zeros that will contain the value of the gradient with respect to theta the displacement phase
+        theta (float): displacement phase
     """
     cutoff = gradTr.shape[0]
     exptheta = np.exp(1j * theta)
@@ -69,7 +73,7 @@ def grad_Dgate(T, gradTr, gradTtheta, theta):  # pragma: no cover
                 gradTr[n, m] -= np.sqrt(m) * T[n, m - 1] * np.conj(exptheta)
 
 
-def Dgate(r, theta, cutoff, grad=False, s=np.arcsinh(1.0)):
+def Dgate(r, theta, cutoff, grad=False, choi_r=np.arcsinh(1.0)):
     """Calculates the Fock representation of the Dgate and its gradient.
 
     Arg:
@@ -77,7 +81,7 @@ def Dgate(r, theta, cutoff, grad=False, s=np.arcsinh(1.0)):
         theta (float): displacement phase
         cutoff (int): Fock ladder cutoff
         grad (boolean): whether to calculate the gradient or not
-        s (float): value of the parameter used internally in fock_tensor
+        choi_r (float): value of the parameter used internally in fock_tensor
 
     Returns:
         tuple[array[complex], array[complex], array[complex]]: The Fock representations of the gate and its gradients with sizes ``[cutoff]*2``
@@ -87,9 +91,9 @@ def Dgate(r, theta, cutoff, grad=False, s=np.arcsinh(1.0)):
     S = np.identity(2 * nmodes)
 
     if not grad:
-        return fock_tensor(S, alpha, cutoff, r=s), None, None
+        return fock_tensor(S, alpha, cutoff, choi_r=choi_r), None, None
 
-    T = np.complex128(fock_tensor(S, alpha, cutoff + 1, r=s))
+    T = np.complex128(fock_tensor(S, alpha, cutoff + 1, choi_r=choi_r))
     gradTr = np.zeros([cutoff, cutoff], dtype=complex)
     gradTtheta = np.zeros([cutoff, cutoff], dtype=complex)
     grad_Dgate(T, gradTr, gradTtheta, theta)
@@ -104,7 +108,7 @@ def grad_Sgate(T, gradTr, gradTtheta, theta):  # pragma: no cover
         T (array[complex]): array representing the gate
         gradTr (array[complex]): array of zeros that will contain the value of the gradient with respect to r
[…]
