+++ thewalrus/tests/test_quantum.py
@@ -646,33 +646,33 @@ def test_total_photon_num_dist_pure_state(cutoff):
     assert np.allclose(p1, p2)
 
 
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_identity(r, tol):
+@pytest.mark.parametrize("choi_r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_identity(choi_r, tol):
     """Tests the correct construction of the single mode identity operation"""
     nmodes = 1
     cutoff = 7
     S = np.identity(2 * nmodes)
     alphas = np.zeros([nmodes])
-    T = fock_tensor(S, alphas, cutoff, r=r)
+    T = fock_tensor(S, alphas, cutoff, choi_r=choi_r)
     expected = np.identity(cutoff)
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_rotation(r, tol):
+@pytest.mark.parametrize("choi_r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_rotation(choi_r, tol):
     """Tests the correct construction of the single mode rotation operation"""
     nmodes = 1
     cutoff = 7
     theta = 2 * np.pi * np.random.rand()
     S = rotation(theta)
     alphas = np.zeros([nmodes])
-    T = fock_tensor(S, alphas, cutoff, r=r)
+    T = fock_tensor(S, alphas, cutoff, choi_r=choi_r)
     expected = np.diag(np.exp(1j * theta * np.arange(cutoff)))
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_displacement(r, tol):
+@pytest.mark.parametrize("choi_r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_displacement(choi_r, tol):
     """Tests the correct construction of the single mode displacement operation"""
     nmodes = 1
     cutoff = 5
@@ -689,12 +689,12 @@ def test_single_mode_displacement(r, tol):
             [-0.01109048 - 1.65323865e-02j, -0.12479885 + 6.30297236e-03j, -0.18306015 + 3.43237787e-01j, 0.27646677 + 4.60777945e-01j, -0.03277289 + 1.88440656e-17j],
         ]
     )
-    T = fock_tensor(S, alphas, cutoff, r=r)
+    T = fock_tensor(S, alphas, cutoff, choi_r=choi_r)
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
-def test_single_mode_squeezing(r, tol):
+@pytest.mark.parametrize("choi_r", [0.5, np.arcsinh(1.0), 2])
+def test_single_mode_squeezing(choi_r, tol):
     """Tests the correct construction of the single mode squeezing operation"""
     nmodes = 1
     s = 1.0
@@ -712,12 +712,12 @@ def test_single_mode_squeezing(r, tol):
             [0.2859358 + 0.0j, 0.0 + 0.0j, -0.29199268 + 0.0j, 0.0 + 0.0j, -0.34474749 + 0.0j],
         ]
     )
-    T = fock_tensor(S, alphas, cutoff, r=r)
+    T = fock_tensor(S, alphas, cutoff, choi_r=choi_r)
     assert np.allclose(T, expected, atol=tol, rtol=0)
 
 
-@pytest.mark.parametrize("r", [0.5, np.arcsinh(1.0), 2])
[…]
