### 2019-12-12 f0b93cd
Updates docs and fock_tensor
