+++ docs/code/fock_gradients.rst
@@ -1,3 +1,4 @@
 .. automodule:: thewalrus.fock_gradients
-	:members:
-	:exclude-members: Dgate, Kgate
+    :members:
+    :inherited-members:
+    :private-members:
