+++ thewalrus/quantum.py
@@ -581,7 +581,7 @@ def find_scaling_adjacency_matrix(A, n_mean):
         n = np.sum(vals2 / (1.0 - vals2))
         return n
 
-    def grad_mean_photon_number(x, vals):
+    def grad_mean_photon_number(x, vals): # pragma: no cover
         r""" Returns the gradient od the mean number of photons in the Gaussian state that
         encodes the adjacency matrix x*A with respect to x.
         vals are the singular values of A
