### 2019-12-13 e63f117
Update thewalrus/quantum.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
