### 2019-12-13 307a687
Update thewalrus/csamples.py
Co-Authored-By: Josh Izaac <josh146@gmail.com>
