+++ thewalrus/csamples.py
@@ -99,6 +99,8 @@ def rescale_adjacency_matrix(
         vals2 = scale * x * vals
         n = np.sum(vals2 / (1.0 - vals2))
         return n
+
+
     # The following function is implicitly tested in test_rescaling_thermal
     def grad_mean_photon_number(x, vals): # pragma: no cover
         r""" Returns the gradient of the mean number of photons in the Gaussian state that
