### 2019-12-19 80c4474
Updates bibliography
