+++ paper.bib
@@ -32,6 +32,16 @@
   doi={10.1016/0304-3975(79)90044-6}
 }
 
+@article{bromley2019,
+  author={{Bromley}, Thomas R. and {Arrazola}, Juan Miguel and {Jahangiri}, Soran and
+         {Izaac}, Josh and {Quesada}, Nicol{\'a}s and {Delgado Gran}, Alain and
+         {Schuld}, Maria and {Swinarton}, Jeremy and {Zabaneh}, Zeid and
+         {Killoran}, Nathan},
+  title="Applications of Near-Term Photonic Quantum Computers: Software and Algorithms",
+  year={2019},
+  journal={arXiv preprint arXiv:1912.07634}
+}
+
 
 @book{ryser1963,
   title={Combinatorial mathematics},
