### 2019-12-19 #1705 @nquesada
It has been a month and half since the peer review ended. Seems like @katyhuff is very busy at the moment. We were wondering if it would be possible to ask another editor to help us finalise the submission process? Maybe @poulson could comment?
