### 2019-12-19 #1705 @danielskatz
Since this seems like it is done, I can take it from here.
