### 2019-12-19 #1705 @danielskatz
I'm also going to proof-read the paper shortly
