### 2019-12-19 #1705 @danielskatz
The paper looks good to me, with the possible exception of some cases in the references.  For example, should "Speedup in classical simulation of gaussian boson sampling" be "Speedup in classical simulation of Gaussian boson sampling"? And should "Franck-condon factors via compressive sensing" be "Franck-Condon factors via compressive sensing"?  Please check cases in the references, and use {}s to protect cases from processing conversion.
