### 2019-12-19 #1705 @nquesada
For the Zenodo submission, is there a quick tutorial on how to do that?
