### 2019-12-19 #1705 @danielskatz
> For the Zenodo submission, is there a quick tutorial on how to do that?

See https://guides.github.com/activities/citable-code/
