### 2019-12-19 #1705 @danielskatz
But you can also tar the repo and deposit in in Zenodo, figshare, or an institutional repository manually
