### 2019-12-19 bf2a3c6
typos fixed
