### 2019-12-19 a72e968
Updates _version.py
