### 2019-12-19 #1705 @josh146
Hi @danielskatz, we've corrected the bibliography per your suggestion, and uploaded to Zenodo (https://zenodo.org/record/3585911#.Xfvb09lyZhE). The DOI is 10.5281/zenodo.3585911. Thanks!
