### 2019-12-19 #1705 @danielskatz
Is the version v0.10.0 ?
